### 2024-09-26 e26d25b
Tentatively fix automatic shutdown.
