+++ constrain/library/automatic_shutdown.py
@@ -9,12 +9,10 @@ class AutomaticShutdown(RuleCheckBase):
         copied_df = (
             self.df.copy()
         )  # copied not to store unnecessary intermediate variables in self.df dataframe
+        copied_df.index.name = "Date"  # rename the index column to Date
         copied_df.reset_index(
             inplace=True
         )  # convert index column back to normal column
-        copied_df = copied_df.rename(
-            columns={"index": "Date"}
-        )  # rename the index column to Date
         copied_df["hvac_set_diff"] = copied_df[
             "hvac_set"
         ].diff()  # calculate the difference between previous and current rows
@@ -29,22 +27,13 @@ class AutomaticShutdown(RuleCheckBase):
             lambda x: x.iloc[[0, -1]]
         )  # group by start/end time
 
-        copied_df["start_time"] = df2["hvac_set_diff"].iloc[::2]  # even number row
-        copied_df["end_time"] = df2["hvac_set_diff"].iloc[1::2]  # odd number row
+        # Get min/max start/end times
+        min_start_time = df2.query("hvac_set_diff == 1")["Date"].dt.hour.min()
+        max_start_time = df2.query("hvac_set_diff == 1")["Date"].dt.hour.max()
+        min_end_time = df2.query("hvac_set_diff == -1")["Date"].dt.hour.min()
+        max_end_time = df2.query("hvac_set_diff == -1")["Date"].dt.hour.max()
 
-        copied_df["min_start_time"] = copied_df.query("start_time == 1")[
-            "Date"
-        ].dt.hour.min()
-        copied_df["max_start_time"] = copied_df.query("start_time == 1")[
-            "Date"
-        ].dt.hour.max()
-        copied_df["min_end_time"] = copied_df.query("end_time == -1")[
-            "Date"
-        ].dt.hour.min()
-        copied_df["max_end_time"] = copied_df.query("end_time == -1")[
-            "Date"
-        ].dt.hour.max()
+        check = (min_start_time != max_start_time) & (min_end_time != max_end_time)
 
-        self.result = (copied_df["min_start_time"] != copied_df["max_start_time"]) & (
-            copied_df["min_end_time"] != copied_df["max_end_time"]
-        )
+        self.df["result"] = check
+        self.result = self.df["result"]
