### 2024-09-30 1052525
add docstring to lib item
