+++ docs/source/conf.py
@@ -29,9 +29,9 @@ copyright = "2023, Pacific Northwest National Laboratory"
 author = "Pacific Northwest National Laboratory"
 
 # The short X.Y version
-version = "0.4.0"
+version = "0.5.0"
 # The full version, including alpha/beta/rc tags
-release = "0.4.0"
+release = "0.5.0"
 
 # Document class docstrings
 autoclass_content = "both"
