### 2024-10-04 eca5a0b
Add functionality to edit state from context menu
