### 2024-10-04 0ab7ca7
Only show edit action if 1 state is selected
