### 2024-10-04 3a5b688
Add more context to GUI
