+++ constrain/app/popup_window.py
@@ -9,19 +9,17 @@ import os
 from PyQt6.QtWidgets import (
     QLineEdit,
     QVBoxLayout,
-    QPushButton,
     QHBoxLayout,
     QComboBox,
     QListWidget,
     QDialog,
     QGroupBox,
     QMessageBox,
-    QSizePolicy,
     QLayout,
 )
-from PyQt6.QtGui import QPixmap
 
 from constrain.app.list_and_choice_popups import ListPopup, ChoicesPopup
+from constrain.app.components.button import StandardButton
 
 script_directory = os.path.dirname(os.path.abspath(__file__))
 dependencies_path = os.path.join(script_directory, "dependencies.json")
@@ -78,13 +76,9 @@ class PopupWindow(QDialog):
         self.form_layout = QVBoxLayout()
 
         self.buttons_layout = QHBoxLayout()
-        self.save_button = QPushButton("Save")
-        self.cancel_button = QPushButton("Cancel")
-        self.payload_button = QPushButton("Edit")
-
-        self.save_button.setFixedSize(100, 20)
-        self.cancel_button.setFixedSize(100, 20)
-        self.payload_button.setFixedSize(100, 20)
+        self.save_button = StandardButton("Save")
+        self.cancel_button = StandardButton("Cancel")
+        self.payload_button = StandardButton("Edit")
 
         self.save_button.clicked.connect(self.save)
         self.cancel_button.clicked.connect(self.close)
@@ -431,8 +425,7 @@ class PopupWindow(QDialog):
         choice_widget = QGroupBox()
         choice_widget.setTitle("Choices")
 
-        edit_button = QPushButton("Edit")
-        edit_button.setFixedSize(100, 20)
+        edit_button = StandardButton("Edit")
         edit_button.clicked.connect(self.choice_popup)
         layout.addWidget(edit_button)
 
@@ -506,8 +499,7 @@ class PopupWindow(QDialog):
         parameter_widget = QGroupBox()
         parameter_widget.setTitle("Parameters")
 
-        parameter_button = QPushButton("Edit")
-        parameter_button.setFixedSize(100, 20)
+        parameter_button = StandardButton("Edit")
         parameter_button.clicked.connect(self.parameter_form)
         layout.addWidget(parameter_button)
 
@@ -564,8 +556,7 @@ class PopupWindow(QDialog):
         payload_widget.setTitle("Payloads")
         layout = QVBoxLayout()
 
-        self.payload_button = QPushButton("Edit")
-        self.payload_button.setFixedSize(100, 20)
+        self.payload_button = StandardButton("Edit")
         self.payload_button.clicked.connect(self.payload_form)
         layout.addWidget(self.payload_button)
 
