### 2024-10-08 11b48f4
Add help tab to toolbar
