### 2024-10-09 d31a537
Add tooltips, fix doc formatting
