+++ tests/test_vav_turndown.py
@@ -11,8 +11,7 @@ from lib_unit_test_runner import *
 class TestVAVTurndown(unittest.TestCase):
     def test_vav_turndown_untested(self):
         points = [
-            "T_AHU_discharge",
-            "T_VAV_discharge",
+            "reheat_coil_flag",
             "V_dot_VAV",
             "V_dot_VAV_max",
             "VAV_min_turndown_design",
@@ -25,8 +24,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [10.0, 10.0, 0.005, 0.01, 0.3, 0.01],
-            [10.0, 10.0, 0.005, 0.01, 0.3, 0.01],
+            [False, 0.005, 0.01, 0.3, 0.01],
+            [False, 0.005, 0.01, 0.3, 0.01],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -42,8 +41,7 @@ class TestVAVTurndown(unittest.TestCase):
 
     def test_vav_turndown_pass(self):
         points = [
-            "T_AHU_discharge",
-            "T_VAV_discharge",
+            "reheat_coil_flag",
             "V_dot_VAV",
             "V_dot_VAV_max",
             "VAV_min_turndown_design",
@@ -56,8 +54,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [55.0, 70.0, 0.005, 0.06, 0.3, 0.01],
-            [55.0, 70.0, 0.005, 0.06, 0.3, 0.01],
+            [True, 0.005, 0.06, 0.3, 0.01],
+            [True, 0.005, 0.06, 0.3, 0.01],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -78,8 +76,7 @@ class TestVAVTurndown(unittest.TestCase):
 
     def test_vav_turndown_fail(self):
         points = [
-            "T_AHU_discharge",
-            "T_VAV_discharge",
+            "reheat_coil_flag",
             "V_dot_VAV",
             "V_dot_VAV_max",
             "VAV_min_turndown_design",
@@ -92,8 +89,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [55.0, 70.0, 0.005, 0.01, 0.3, 0.01],
-            [55.0, 70.0, 0.005, 0.01, 0.3, 0.01],
+            [True, 0.005, 0.01, 0.3, 0.01],
+            [True, 0.005, 0.01, 0.3, 0.01],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
