### 2024-10-29 36452a8
Added Untested case when V_dot_VAV_max = zero
