### 2024-11-05 0a1a04e
developed verification lib
