### 2024-11-05 0653a7d
Added doc string, Updated library
