### 2024-11-06 2cabe42
Simplified check_bool method logic
