+++ tests/test_vav_minimum_turndown_during_reheat.py
@@ -30,7 +30,7 @@ class TestVAVTurndown(unittest.TestCase):
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
-        verification_obj = run_test_verification_with_data("VAVTurndown", df)
+        verification_obj = run_test_verification_with_data("VAVMinimumTurndownDuringReheat", df)
         results = list(verification_obj.result)
         expected_results = [
             "Untested",
@@ -60,7 +60,7 @@ class TestVAVTurndown(unittest.TestCase):
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
-        verification_obj = run_test_verification_with_data("VAVTurndown", df)
+        verification_obj = run_test_verification_with_data("VAVMinimumTurndownDuringReheat", df)
         results = list(verification_obj.result)
         expected_results = [
             "Untested",
@@ -90,7 +90,7 @@ class TestVAVTurndown(unittest.TestCase):
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
-        verification_obj = run_test_verification_with_data("VAVTurndown", df)
+        verification_obj = run_test_verification_with_data("VAVMinimumTurndownDuringReheat", df)
 
         results = pd.Series(list(verification_obj.result))
         expected_results = pd.Series(
@@ -125,7 +125,7 @@ class TestVAVTurndown(unittest.TestCase):
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
-        verification_obj = run_test_verification_with_data("VAVTurndown", df)
+        verification_obj = run_test_verification_with_data("VAVMinimumTurndownDuringReheat", df)
 
         results = pd.Series(list(verification_obj.result))
         expected_results = pd.Series(
