### 2024-11-06 0015671
Black formatting
