### 2024-11-07 1c36156
Fixed plotting error when Untested results exist, Added Untested column in report
