+++ tests/test_vav_turndown_during_reheat.py
@@ -126,3 +126,4 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         self.assertEqual(results, expected_results)
+        self.assertEqual(verification_obj.check_bool(), "Untested")
