### 2024-11-08 a91e8ea
Added check_bool checking part in unit test
