+++ constrain/library/vav_turndown_during_reheat.py
@@ -49,8 +49,8 @@ class VAVTurndownDuringReheat(RuleCheckBase):
             self.df["V_dot_VAV_max"] > 0
         ).all(), "Not all `V_dot_VAV_max` values are greater than 0"
 
-        # Check if the `reheat_coil_flag` column has only False values
-        if (self.df["reheat_coil_flag"] == False).all():
+        # Check if the `reheat_coil_flag` column has only True/False values
+        if self.df["reheat_coil_flag"].nunique() == 1:
             self.df["result"] = "Untested"
         else:
             self.df["V_dot_VAV_ratio"] = self.df["V_dot_VAV"] / self.df["V_dot_VAV_max"]
