+++ tests/test_vav_turndown_during_reheat.py
@@ -91,7 +91,7 @@ class TestVAVTurndown(unittest.TestCase):
         binary_result = verification_obj.check_bool()
         self.assertFalse(binary_result)
 
-    def test_vav_turndown_during_reheat_untested(self):
+    def test_vav_turndown_during_reheat_no_reheat_untested(self):
         points = [
             "reheat_coil_flag",
             "V_dot_VAV",
@@ -110,6 +110,43 @@ class TestVAVTurndown(unittest.TestCase):
             [False, 370, 620],
             [False, 360, 620],
             [False, 380, 620],
+        ]
+
+        df = pd.DataFrame(data, columns=points, index=timestamp)
+
+        verification_obj = run_test_verification_with_data(
+            "VAVTurndownDuringReheat", df
+        )
+        results = list(verification_obj.result)
+        expected_results = [
+            "Untested",
+            "Untested",
+            "Untested",
+            "Untested",
+        ]
+
+        self.assertEqual(results, expected_results)
+        self.assertEqual(verification_obj.check_bool(), "Untested")
+
+    def test_vav_turndown_during_reheat_all_reheat_untested(self):
+        points = [
+            "reheat_coil_flag",
+            "V_dot_VAV",
+            "V_dot_VAV_max",
+        ]
+
+        timestamp = [
+            datetime(2024, 8, 1, 12, 0, 0),
+            datetime(2024, 8, 1, 13, 0, 0),
+            datetime(2024, 8, 1, 14, 0, 0),
+            datetime(2024, 8, 1, 15, 0, 0),
+        ]
+
+        data = [
+            [True, 350, 620],
+            [True, 370, 620],
+            [True, 360, 620],
+            [True, 380, 620],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
