+++ constrain/api/verification.py
@@ -48,17 +48,19 @@ class Verification:
     def configure(
         self,
         output_path: str = None,
+        time_series_csv_export_name: str = None,
         lib_items_path: str = None,
         lib_classes_py_file: str = None,
         plot_option: str = None,
         fig_size: tuple = (6.4, 4.8),
-        num_threads: int = 1,
+        num_threads: int = 1,  # currently useless TODO: fix this
         preprocessed_data: pd.DataFrame = None,
     ) -> None:
         """Configure verification environment.
 
         Args:
             output_path (str): Verification results output path.
+            time_series_archive_name (str, optional): CSV file name for saving a complete data csv file with verification result flags. Defaults to None, which will not save any time series data archives.
             lib_items_path (str, optional): User provided verification item json path (include name of the file with extension).
             lib_classes_py_file (str, optional): User provided verification item python classes file.
             plot_option (str, optional): Type of plots to include. It should either be all-compact, all-expand, day-compact, or day-expand. It can also be None, which will plot all types. Default to None.
@@ -78,6 +80,13 @@ class Verification:
             logging.error("The specificed output directory does not exist.")
             return None
 
+        if time_series_csv_export_name is not None:
+            if not isinstance(time_series_csv_export_name, str):
+                logging.error("time_series_csv_export_name should be a string.")
+                return None
+
+        # TODO: lib_items_path now only needed when user provides their own lib items, and the default lib items from
+        #  ConStrain will be loaded without user inputs. This is no longer an error to be logged.
         if lib_items_path is None:
             logging.error(
                 "A path to the library of verification cases should be provided."
@@ -136,6 +145,11 @@ class Verification:
             return None
 
         self.output_path = output_path
+
+        self.time_series_csv_export_name = time_series_csv_export_name
+        if time_series_csv_export_name[-4:] == ".csv":
+            self.time_series_csv_export_name = time_series_csv_export_name[:-4]
+
         self.lib_items_path = lib_items_path
         self.lib_classes_py_file = lib_classes_py_file
         self.plot_option = plot_option
@@ -167,6 +181,7 @@ class Verification:
             user_lib_file=self.lib_classes_py_file,
             plot_option=self.plot_option,
             output_path=self.output_path,
+            time_series_file_name=self.time_series_csv_export_name,
             fig_size=self.fig_size,
             produce_outputs=True,
             preprocessed_data=self.preprocessed_data,
