### 2024-11-19 50abe47
add verification results to verification df
