### 2024-11-19 c58c290
replace nan with untested
