+++ tests/api/data/custom_lib/UserProvidedVerificationItem1.py
@@ -13,7 +13,7 @@ class UserProvidedVerificationItem1(RuleCheckBase):
             else:
                 return True
         else:
-            return np.nan
+            return "Untested"
 
     def verify(self):
         self.result = self.df.apply(lambda d: self.automatic_oa_damper_check(d), axis=1)
