### 2024-11-21 d211413
add pass fail backgroun color to plotly
