### 2024-12-05 4ea59a7
add plotly to dependency
