### 2024-12-05 786f06d
accomodate untested in interactive plot
