+++ constrain/checklib.py
@@ -285,26 +285,27 @@ class CheckLibBase(ABC):
         plotday_filtered = plotday[plotday != "Untested"]
 
         # flag
-        ax1 = plt.subplot(2, 1, 1)
-        sns.scatterplot(x=plotday_filtered.index, y=plotday_filtered)
-        plt.xlim([plotday.index[0], plotday.index[-1]])
-        plt.ylim([-0.2, 1.2])
-        plt.title(f"Example day Pass / Fail flag - {self.__class__.__name__}")
+        if plotday_filtered.size != 0:
+            ax1 = plt.subplot(2, 1, 1)
+            sns.scatterplot(x=plotday_filtered.index, y=plotday_filtered)
+            plt.xlim([plotday.index[0], plotday.index[-1]])
+            plt.ylim([-0.2, 1.2])
+            plt.title(f"Example day Pass / Fail flag - {self.__class__.__name__}")
 
-        # datapoints
-        ax2 = plt.subplot(2, 1, 2)
-        plotdaydf[plt_pts].plot(ax=ax2)
-        pt_nan = plotdaydf.isnull().any().to_dict()
-        for i, line in enumerate(ax2.get_lines()):
-            line_label = line.get_label()
-            if pt_nan[line_label]:
-                line.set_marker(".")
-        ax2.ticklabel_format(useOffset=False, axis="y")
+            # datapoints
+            ax2 = plt.subplot(2, 1, 2)
+            plotdaydf[plt_pts].plot(ax=ax2)
+            pt_nan = plotdaydf.isnull().any().to_dict()
+            for i, line in enumerate(ax2.get_lines()):
+                line_label = line.get_label()
+                if pt_nan[line_label]:
+                    line.set_marker(".")
+            ax2.ticklabel_format(useOffset=False, axis="y")
 
-        plt.title(f"Example day data points plot - {self.__class__.__name__}")
-        plt.tight_layout()
-        plt.savefig(f"{self.results_folder}/Day_plot_aio.png")
-        print()
+            plt.title(f"Example day data points plot - {self.__class__.__name__}")
+            plt.tight_layout()
+            plt.savefig(f"{self.results_folder}/Day_plot_aio.png")
+            print()
 
     def day_plot_obo(self, plt_pts, fig_size):
         """One by one plot of all samples"""
@@ -314,33 +315,34 @@ class CheckLibBase(ABC):
         plotday, plotdaydf = self.calculate_plot_day()
         plotday_filtered = plotday[plotday != "Untested"]
         # flag
-        ax1 = plt.subplot(num_plots, 1, 1)
-        sns.scatterplot(x=plotday_filtered.index, y=plotday_filtered)
-        plt.xlim([plotday.index[0], plotday.index[-1]])
-        plt.ylim([-0.2, 1.2])
-        plt.title(f"Example day Pass / Fail flag plot - {self.__class__.__name__}")
+        if plotday_filtered.size != 0:
+            ax1 = plt.subplot(num_plots, 1, 1)
+            sns.scatterplot(x=plotday_filtered.index, y=plotday_filtered)
+            plt.xlim([plotday.index[0], plotday.index[-1]])
+            plt.ylim([-0.2, 1.2])
+            plt.title(f"Example day Pass / Fail flag plot - {self.__class__.__name__}")
 
-        # datapoints
-        pt_nan = plotdaydf.isnull().any().to_dict()
-        i = 2
-        for pt in plt_pts:
-            try:
-                axx = plt.subplot(num_plots, 1, i)
-                if pt_nan[pt]:
-                    plotdaydf[pt].plot(ax=axx, marker=".")
-                else:
-                    # check if values in the series are boolean
[…]
