### 2024-12-05 6a82de5
Added unit test for untested case
