### 2024-12-05 3ea7a05
limit flag shading
