### 2024-12-06 154961c
Added V_dot_ratio in the dataframe
