+++ constrain/library/vav_minimum_turndown_during_reheat.py
@@ -60,4 +60,8 @@ class VAVMinimumTurndownDuringReheat(RuleCheckBase):
             return "Untested"
 
     def verify(self):
+        if (self.df["V_dot_VAV_max"] != 0).all():
+            self.df["V_dot_ratio"] = (
+                self.df["V_dot_VAV"] / self.df["V_dot_VAV_max"]
+            )  # for plotting
         self.result = self.df.apply(lambda d: self.vav_turndown_check(d), axis=1)
