### 2024-12-06 81e2634
Reverted changes in checklib.py
