### 2024-12-30 6f00a5b
Fix verification logic
