### 2025-01-23 #7195 @FWuellhorst
@AdamRJensen As far as I can see, the framework relies on ASHRAE specifications on how controls should work in the US. As I am from Europe and have no knowledge on these guidelines, it would be hard for me to review whether the framework solves the statement of need.
