### 2025-01-28 #7195 @AdamRJensen
> [@AdamRJensen](https://github.com/AdamRJensen) As far as I can see, the framework relies on ASHRAE specifications on how controls should work in the US. As I am from Europe and have no knowledge on these guidelines, it would be hard for me to review whether the framework solves the statement of need.

Hi @FWuellhorst thanks for bringing this up. The review is more about the software than the very specific guidelines, so if you have experience on software and building system controls then you would be a great fit. While ideal, I find it very unlikely that we will find an ASHRAE expert with software skills for this review.
