### 2025-03-13 #7195 @AdamRJensen
> @AdamRJensen , we are having trouble identifying additional reviewers from https://reviewers.joss.theoj.org/reviewers/search, is it acceptable to have reviewers from outside of that pool?

@leijerry888 most certainly!
