### 2025-03-20 e8a9a5c
Reactivate multiprocessing; Move folders so example can be ran outside of a clone.
