+++ constrain/examples.py
@@ -4,15 +4,17 @@ examples.py
 This modules provide helper functions to retrieve data and information for examples of ConStrain.
 """
 
-import logging
+import logging, pathlib
 from .api import DataProcessing
+
+path = pathlib.Path(__file__).parent.resolve()
 
 # TODO: move this to a JSON file
 examples = {
     "example_1": {
         "description": "Perform verification of ASHRAE Guideline 36-2021 sequence of operation on a dataset generated through the simulation of an AHU in Modelica. The verifications include the following: supply temperature reset, outdoor air damper psition for relief damper/fan, and return air damper psition for relief damper/fan",
-        "path_to_data": "./demo/G36_demo/data/G36_Modelica_Jan.csv",
-        "path_to_verifications": "./demo/G36_demo/data/G36_library_verification_cases.json",
+        "path_to_data": f"{path}/demo/G36_demo/data/G36_Modelica_Jan.csv",
+        "path_to_verifications": f"{path}/demo/G36_demo/data/G36_library_verification_cases.json",
     }
 }
 
@@ -40,7 +42,7 @@ class Examples:
             return
 
     def library(self):
-        return "./schema/library.json"
+        return f"{path}/schema/library.json"
 
     def verifications(self, example_name):
         if self.check_example(example_name):
