### 2025-03-21 f06ca1d
Clarify point labels.
