+++ constrain/library_outdated.py
@@ -42,14 +42,14 @@ class IntegratedEconomizerControl(CheckLibBase):
 
 
 class SupplyAirTempReset(RuleCheckBase):
-    points = ["T_sa_set", "T_z_coo"]
+    points = ["T_sa_set", "T_z_cool"]
 
     def verify(self):
         t_sa_set_max = max(self.df["T_sa_set"])
         t_sa_set_min = min(self.df["T_sa_set"])
 
         self.result = (t_sa_set_max - t_sa_set_min) >= (
-            self.df["T_z_coo"] - t_sa_set_min
+            self.df["T_z_cool"] - t_sa_set_min
         ) * 0.25 * 0.99  # 0.99 being the numeric threshold
 
     def plot(self, plot_option, plt_pts=None):
@@ -84,12 +84,12 @@ class EconomizerHighLimitA(RuleCheckBase):
 
 
 class EconomizerHighLimitB(RuleCheckBase):
-    points = ["oa_flow", "T_oa_db", "ret_a_temp", "oa_min_flow"]
+    points = ["oa_flow", "T_oa_db", "T_ra", "oa_min_flow"]
 
     def verify(self):
         self.result = ~(
             (self.df["oa_flow"] > self.df["oa_min_flow"])
-            & (self.df["ret_a_temp"] < self.df["T_oa_db"])
+            & (self.df["T_ra"] < self.df["T_oa_db"])
         )
 
 
@@ -134,10 +134,10 @@ class EconomizerHighLimitD(RuleCheckBase):
 
 
 class ZoneTempControl(RuleCheckBase):
-    points = ["T_set_cool", "T_set_heat"]
+    points = ["T_z_cool_sp", "T_z_heat_sp"]
 
     def verify(self):
-        self.result = (self.df["T_set_cool"] - self.df["T_set_heat"]) > 2.77
+        self.result = (self.df["T_z_cool_sp"] - self.df["T_z_heat_sp"]) > 2.77
 
 
 class HWReset(RuleCheckBase):
@@ -235,10 +235,10 @@ class CHWReset(RuleCheckBase):
 
 
 class ZoneHeatSetpointMinimum(CheckLibBase):
-    points = ["T_heat_set"]
+    points = ["T_z_heat_sp"]
 
     def verify(self):
-        self.result = self.df["T_heat_set"] <= 12.78
+        self.result = self.df["T_z_heat_sp"] <= 12.78
 
     def check_bool(self) -> bool:
         if len(self.result[self.result == True] > 0):
@@ -260,10 +260,10 @@ class ZoneHeatSetpointMinimum(CheckLibBase):
 
 
 class ZoneCoolingSetpointMaximum(CheckLibBase):
-    points = ["T_cool_set"]
+    points = ["T_z_cool_sp"]
 
     def verify(self):
-        self.result = self.df["T_cool_set"] >= 32.22
+        self.result = self.df["T_z_cool_sp"] >= 32.22
 
     def check_bool(self) -> bool:
         if len(self.result[self.result == True] > 0):
@@ -285,13 +285,13 @@ class ZoneCoolingSetpointMaximum(CheckLibBase):
 
 
 class ZoneHeatingResetDepth(CheckLibBase):
-    points = ["T_heat_set"]
+    points = ["T_z_heat_sp"]
 
     def verify(self):
-        self.df["t_heat_set_min"] = min(self.df["T_heat_set"])
-        self.df["t_heat_set_max"] = max(self.df["T_heat_set"])
+        self.df["T_z_heat_sp_min"] = min(self.df["T_z_heat_sp"])
+        self.df["T_z_heat_sp_max"] = max(self.df["T_z_heat_sp"])
 
[…]
