### 2025-03-21 1ecc46c
Remove unwanted print statement.
