+++ constrain/library/G36ReheatTerminalBoxCoolingAirflowSetpoint.py
@@ -45,7 +45,6 @@ end
 """
 
 from constrain.checklib import RuleCheckBase
-import numpy as np
 
 
 class G36ReheatTerminalBoxCoolingAirflowSetpoint(RuleCheckBase):
