+++ constrain/library/G36CoolingOnlyTerminalBoxCoolingAirflowSetpoint.py
@@ -42,7 +42,6 @@ end
 """
 
 from constrain.checklib import RuleCheckBase
-import numpy as np
 
 
 class G36CoolingOnlyTerminalBoxCoolingAirflowSetpoint(RuleCheckBase):
