+++ constrain/library/demand_control_vent.py
@@ -1,6 +1,4 @@
-import numpy as np
 import pandas as pd
-
 from constrain.checklib import CheckLibBase
 from scipy.stats import pearsonr
 
