+++ constrain/library/G36CoolingOnlyTerminalBoxDeadbandAirflowSetpoint.py
@@ -37,7 +37,6 @@ end
 """
 
 from constrain.checklib import RuleCheckBase
-import numpy as np
 
 
 class G36CoolingOnlyTerminalBoxDeadbandAirflowSetpoint(RuleCheckBase):
