+++ constrain/library/G36TerminalBoxCoolingMinimumAirflow.py
@@ -36,7 +36,6 @@ else:
 """
 
 from constrain.checklib import RuleCheckBase
-import numpy as np
 
 
 class G36TerminalBoxCoolingMinimumAirflow(RuleCheckBase):
