### 2025-03-24 906db4f
Updated automatic_shutdown docstring
