### 2025-03-24 1cd936e
Add general tolerances instead of having them being specified for each verification case.
