+++ tests/test_lib_automatic_oa_damper_controls.py
@@ -11,21 +11,31 @@ import numpy as np
 
 class TestAutomaticOADamperControl(unittest.TestCase):
     def test_automatic_oa_damper_control(self):
-        points = ["o", "eco_onoff", "m_oa", "m_ea", "tol_o", "tol_m_oa", "tol_m_ea"]
+        tolerances = {
+            "airflow": {
+                "unit": "m3/s",
+                "general": 50,
+                "types": {"exhaust_air": 50, "outdoor_air": 50},
+            },
+            "ratio": {"unit": "%", "general": 0.001, "types": {"occupancy": 0.001}},
+        }
+        points = ["o", "eco_onoff", "m_oa", "m_ea"]
         data = [
-            [1, 0, 0, 0, 0.001, 50, 50],
-            [0, 1, 0, 0, 0.001, 50, 50],
-            [0, 0, 1000, 0, 0.001, 50, 50],
-            [0, 0, 0, 1000, 0.001, 50, 50],
-            [0, 0, 1000, 1000, 0.001, 50, 50],
-            [1, 1, 0, 0, 0.001, 50, 50],
-            [1, 0, 1, 0, 0.001, 50, 50],
-            [1, 1, 1000, 0, 0.001, 50, 50],
+            [1, 0, 0, 0],
+            [0, 1, 0, 0],
+            [0, 0, 1000],
+            [0, 0, 0, 1000],
+            [0, 0, 1000, 1000],
+            [1, 1, 0, 0],
+            [1, 0, 1, 0],
+            [1, 1, 1000, 0],
         ]
         df = pd.DataFrame(data, columns=points)
 
         results = list(
-            run_test_verification_with_data("AutomaticOADamperControl", df).result
+            run_test_verification_with_data(
+                "AutomaticOADamperControl", df, tolerances=tolerances
+            ).result
         )
         expected_results = [
             "Untested",
