### 2025-03-24 2e476ff
Back-out changes from the clean_point_labels branch.
