### 2025-03-24 29a4ec5
Updated docstrings in the verification items
