+++ constrain/library_outdated.py
@@ -134,10 +134,10 @@ class EconomizerHighLimitD(RuleCheckBase):
 
 
 class ZoneTempControl(RuleCheckBase):
-    points = ["T_set_cool", "T_set_heat"]
+    points = ["T_cool_set", "T_heat_set"]
 
     def verify(self):
-        self.result = (self.df["T_set_cool"] - self.df["T_set_heat"]) > 2.77
+        self.result = (self.df["T_cool_set"] - self.df["T_heat_set"]) > 2.77
 
 
 class HWReset(RuleCheckBase):
@@ -235,10 +235,10 @@ class CHWReset(RuleCheckBase):
 
 
 class ZoneHeatSetpointMinimum(CheckLibBase):
-    points = ["T_set_heat"]
+    points = ["T_heat_set"]
 
     def verify(self):
-        self.result = self.df["T_set_heat"] <= 12.78
+        self.result = self.df["T_heat_set"] <= 12.78
 
     def check_bool(self) -> bool:
         if len(self.result[self.result == True] > 0):
@@ -260,10 +260,10 @@ class ZoneHeatSetpointMinimum(CheckLibBase):
 
 
 class ZoneCoolingSetpointMaximum(CheckLibBase):
-    points = ["T_set_cool"]
+    points = ["T_cool_set"]
 
     def verify(self):
-        self.result = self.df["T_set_cool"] >= 32.22
+        self.result = self.df["T_cool_set"] >= 32.22
 
     def check_bool(self) -> bool:
         if len(self.result[self.result == True] > 0):
@@ -285,13 +285,13 @@ class ZoneCoolingSetpointMaximum(CheckLibBase):
 
 
 class ZoneHeatingResetDepth(CheckLibBase):
-    points = ["T_set_heat"]
+    points = ["T_heat_set"]
 
     def verify(self):
-        self.df["T_set_heat_min"] = min(self.df["T_set_heat"])
-        self.df["T_set_heat_max"] = max(self.df["T_set_heat"])
+        self.df["T_heat_set_min"] = min(self.df["T_heat_set"])
+        self.df["T_heat_set_max"] = max(self.df["T_heat_set"])
 
-        self.result = (self.df["T_set_heat_max"] - self.df["T_set_heat_min"]) >= 5.55
+        self.result = (self.df["T_heat_set_max"] - self.df["T_heat_set_min"]) >= 5.55
 
     def check_bool(self) -> bool:
         if len(self.result[self.result == True] > 0):
@@ -305,8 +305,8 @@ class ZoneHeatingResetDepth(CheckLibBase):
             "Pass #": len(self.result[self.result == True]),
             "Fail #": len(self.result[self.result == False]),
             "Verification Passed?": self.check_bool(),
-            "max(T_set_heat)": self.df["T_set_heat_max"][0],
-            "min(T_set_heat)": self.df["T_set_heat_min"][0],
+            "max(T_heat_set)": self.df["T_heat_set_max"][0],
+            "min(T_heat_set)": self.df["T_heat_set_min"][0],
         }
 
         print("Verification results dict: ")
@@ -315,13 +315,13 @@ class ZoneHeatingResetDepth(CheckLibBase):
 
 
 class ZoneCoolingResetDepth(CheckLibBase):
-    points = ["T_set_cool"]
+    points = ["T_cool_set"]
 
     def verify(self):
-        self.df["T_set_cool_min"] = min(self.df["T_set_cool"])
-        self.df["T_set_cool_max"] = max(self.df["T_set_cool"])
+        self.df["T_cool_set_min"] = min(self.df["T_cool_set"])
+        self.df["T_cool_set_max"] = max(self.df["T_cool_set"])
 
[…]
