+++ constrain/library/guest_room_control_vent.py
@@ -1,4 +1,4 @@
-from constrain.checklib import CheckLibBase
+from constrain.checklib import RuleCheckBase
 import pandas as pd
 
 
