### 2025-03-25 060e434
Remove tolerance from plotting.
