### 2025-03-25 33bd7d4
Remove tolerance and revert formatting changes.
