### 2025-03-25 2419517
Clarify tolerance usage and fix error in the library.json file.
