+++ constrain/library/MZSystemOccupiedStandbyVentilationZoneControl.py
@@ -57,7 +57,7 @@ class MZSystemOccupiedStandbyVentilationZoneControl(RuleCheckBase):
                 - data["m_oa_requested_by_system"]
             ) >= (
                 data["m_oa_zone_requirement"]
-                * (1 - self.get_tolerance("airflow", "outdoor_air"))
+                - self.get_tolerance("airflow", "outdoor_air")
             ):
                 return True
             else:
