### 2025-03-25 9342751
Clarify tolerance lookup and fix maximum command values.
