### 2025-03-25 1e005c7
Add exterior lighting tolerance.
