### 2025-03-25 71d605c
Made docstrings aligned with library.json
