+++ constrain/library/automatic_shutdown.py
@@ -37,7 +37,7 @@ end
 
 ### Data requirements
 
--  flag_hvac: HVAC operation schedule
+-  schedule_hvac: HVAC operation schedule
   - Data Value Unit: binary
   - Data point Description: HVAC system status
   - Data Point Affiliation: HVAC operation schedule
@@ -49,7 +49,7 @@ from constrain.checklib import RuleCheckBase
 
 
 class AutomaticShutdown(RuleCheckBase):
-    points = ["hvac_set"]
+    points = ["schedule_hvac"]
 
     def verify(self):
         copied_df = (
@@ -59,12 +59,12 @@ class AutomaticShutdown(RuleCheckBase):
         copied_df.reset_index(
             inplace=True
         )  # convert index column back to normal column
-        copied_df["hvac_set_diff"] = copied_df[
-            "hvac_set"
+        copied_df["hvac_operation_diff"] = copied_df[
+            "schedule_hvac"
         ].diff()  # calculate the difference between previous and current rows
         copied_df = copied_df.dropna(axis=0)  # drop NaN row
         copied_df = copied_df.loc[
-            copied_df["hvac_set_diff"] != 0.0
+            copied_df["hvac_operation_diff"] != 0.0
         ]  # filter out 0.0 values
         copied_df["Date"] = pd.to_datetime(
             copied_df["Date"], format="%Y-%m-%d %H:%M:%S"
@@ -74,10 +74,10 @@ class AutomaticShutdown(RuleCheckBase):
         )  # group by start/end time
 
         # Get min/max start/end times
-        min_start_time = df2.query("hvac_set_diff == 1")["Date"].dt.hour.min()
-        max_start_time = df2.query("hvac_set_diff == 1")["Date"].dt.hour.max()
-        min_end_time = df2.query("hvac_set_diff == -1")["Date"].dt.hour.min()
-        max_end_time = df2.query("hvac_set_diff == -1")["Date"].dt.hour.max()
+        min_start_time = df2.query("hvac_operation_diff == 1")["Date"].dt.hour.min()
+        max_start_time = df2.query("hvac_operation_diff == 1")["Date"].dt.hour.max()
+        min_end_time = df2.query("hvac_operation_diff == -1")["Date"].dt.hour.min()
+        max_end_time = df2.query("hvac_operation_diff == -1")["Date"].dt.hour.max()
 
         check = (min_start_time != max_start_time) & (min_end_time != max_end_time)
 

[…]
