+++ tests/test_g36_OutdoorAirDamperPositionForReturnFanDirectBuildingPressure.py
@@ -11,9 +11,9 @@ class TestG36OutdoorAirDamperPositionForReturnFanDirectBuildingPressure(
 ):
     def test_return_air_damper_position(self):
         points = [
-            "oa_p",
-            "max_oa_p",
-            "oa_p_tol",
+            "cmd_damper_oa",
+            "cmd_damper_oa_max",
+            "tol_cmd_damper_oa",
         ]
         data = [
             [0.1, 1, 0.05],  # False
