+++ tests/test_fan_static_pressure_reset_control.py
@@ -11,12 +11,12 @@ from lib_unit_test_runner import *
 class TestFanStaticPressureResetControl(unittest.TestCase):
     def test_fan_static_pressure_reset_control_pass(self):
         points = [
-            "p_set",
-            "d_VAV_1",
-            "d_VAV_2",
-            "d_VAV_3",
-            "d_VAV_4",
-            "d_VAV_5",
+            "p_press_static_sp",
+            "pos_damper_vav_1",
+            "pos_damper_vav_2",
+            "pos_damper_vav_3",
+            "pos_damper_vav_4",
+            "pos_damper_vav_5",
         ]
 
         timestamp = [
@@ -55,12 +55,12 @@ class TestFanStaticPressureResetControl(unittest.TestCase):
 
     def test_fan_static_pressure_reset_control_fail(self):
         points = [
-            "p_set",
-            "d_VAV_1",
-            "d_VAV_2",
-            "d_VAV_3",
-            "d_VAV_4",
-            "d_VAV_5",
+            "p_press_static_sp",
+            "pos_damper_vav_1",
+            "pos_damper_vav_2",
+            "pos_damper_vav_3",
+            "pos_damper_vav_4",
+            "pos_damper_vav_5",
         ]
 
         timestamp = [
