### 2025-03-25 f1e79e5
Updated TestAppendixGHVACSystemFanOperation unit test
