+++ tests/test_appendixg_HVACSystemFanOperation.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 class TestAppendixGHVACSystemFanOperation(unittest.TestCase):
     def test_hvac_system_fan_operation_no_oa_untested(self):
-        points = ["o", "fan_runtime_fraction", "m_oa", "tol_o"]
+        points = ["n_occupants", "frac_runtime_fan", "v_oa", "tol_occupants"]
         timestamp = [
             datetime(2023, 3, 1, 2, 0, 0),
             datetime(2023, 3, 1, 2, 5, 2),
@@ -26,7 +26,7 @@ class TestAppendixGHVACSystemFanOperation(unittest.TestCase):
         self.assertTrue(binaryflag is None)
 
     def test_hvac_system_fan_operation_oa_fail(self):
-        points = ["o", "fan_runtime_fraction", "m_oa", "tol_o"]
+        points = ["n_occupants", "frac_runtime_fan", "v_oa", "tol_occupants"]
         timestamp = [
             datetime(2023, 3, 1, 2, 0, 0),
             datetime(2023, 3, 1, 2, 5, 2),
@@ -48,7 +48,7 @@ class TestAppendixGHVACSystemFanOperation(unittest.TestCase):
         self.assertFalse(binaryflag)
 
     def test_hvac_system_fan_operation_oa_pass(self):
-        points = ["o", "fan_runtime_fraction", "m_oa", "tol_o"]
+        points = ["n_occupants", "frac_runtime_fan", "v_oa", "tol_occupants"]
         timestamp = [
             datetime(2023, 3, 1, 2, 0, 0),
             datetime(2023, 3, 1, 2, 5, 2),
