+++ tests/api/data/custom_lib_aio/custom_lib_all.py
@@ -345,12 +345,12 @@ from constrain import CheckLibBase, RuleCheckBase
 
 
 class UserProvidedVerificationItem1(RuleCheckBase):
-    points = ["o", "eco_onoff", "m_oa", "m_ea", "tol_o", "tol_m_oa", "tol_m_ea"]
+    points = ["n_occupants", "status_economizer", "v_oa", "v_ea", "tol_occupants", "tol_v_oa", "tol_v_ea"]
 
     def automatic_oa_damper_check(self, data):
-        if data["o"] < data["tol_o"]:
-            if data["eco_onoff"] == 0 and (
-                data["m_oa"] >= data["tol_m_oa"] or data["m_ea"] >= data["tol_m_ea"]
+        if data["n_occupants"] < data["tol_occupants"]:
+            if data["status_economizer"] == 0 and (
+                data["v_oa"] >= data["tol_v_oa"] or data["v_ea"] >= data["tol_v_ea"]
             ):
                 return False
             else:
@@ -363,12 +363,12 @@ class UserProvidedVerificationItem1(RuleCheckBase):
 
 
 class UserProvidedVerificationItem_Beta(RuleCheckBase):
-    points = ["o", "eco_onoff", "m_oa", "m_ea", "tol_o", "tol_m_oa", "tol_m_ea"]
+    points = ["n_occupants", "status_economizer", "v_oa", "v_ea", "tol_occupants", "tol_v_oa", "tol_v_ea"]
 
     def automatic_oa_damper_check(self, data):
-        if data["o"] < data["tol_o"]:
-            if data["eco_onoff"] == 0 and (
-                data["m_oa"] >= data["tol_m_oa"] or data["m_ea"] >= data["tol_m_ea"]
+        if data["n_occupants"] < data["tol_occupants"]:
+            if data["status_economizer"] == 0 and (
+                data["v_oa"] >= data["tol_v_oa"] or data["v_ea"] >= data["tol_v_ea"]
             ):
                 return False
             else:

[…]
