+++ constrain/checklib.py
@@ -153,16 +153,8 @@ class CheckLibBase(ABC):
         """
         if self.tolerances is not None:
             if variable_type in self.tolerances.keys():
-                if variable_subtype == "general":
-                    return self.tolerances[variable_type]["general"]
-                else:
-                    if (
-                        variable_subtype
-                        in self.tolerances[variable_type]["types"].keys()
-                    ):
-                        return self.tolerances[variable_type]["types"][variable_subtype]
-                    else:
-                        return 0.0
+                print(self.tolerances[variable_type])
+                return self.tolerances[variable_type]["types"][variable_subtype]
             else:
                 return 0.0
         else:
