+++ tests/test_g36_ReliefDamperStatus.py
@@ -14,10 +14,10 @@ class TestG36ReliefDamperStatus(unittest.TestCase):
     tolerances = {
         "damper": {
             "unit": "%",
-            "general": 0.0,
             "types": {
                 "position": 0.0,
                 "command": 0.0,
+                "general": 0.0,
             },
         }
     }
