+++ tests/api/test_verification.py
@@ -324,9 +324,9 @@ class TestVerification(unittest.TestCase):
                 fig_size=(6, 5),
                 num_threads=1,
                 preprocessed_data=df.data,
-                path_to_custom_tolerance_file="../constrain/tolerances.json",
+                path_to_custom_tolerance_file="./constrain/tolerances.json",
             )
-            assert len(logobs.output) == 12
+            assert len(logobs.output) == 11
 
     def test_run_single_verification(self):
         # Single verification
