### 2025-03-27 79a3368
Add tolerance for coil load.
