+++ tests/test_g36_OutdoorAirDamperPositionForReliefDamperOrFan.py
@@ -9,14 +9,14 @@ import pandas as pd
 class TestG36OutdoorAirDamperPositionForReliefDamperOrFan(unittest.TestCase):
     def test_return_air_damper_position(self):
         points = [
-            "q_heat",
-            "q_cool",
-            "pos_damper_ra",
-            "pos_damper_ra_max",
+            "output_coil_heating",
+            "output_coil_cooling",
+            "position_damper_air_return",
+            "position_damper_air_return_max",
             "tol_pos_damper_ra",
-            "pos_damper_oa",
-            "pos_damper_oa_min",
-            "pos_damper_oa_max",
+            "position_damper_air_outdoor",
+            "position_damper_air_outdoor_min",
+            "position_damper_air_outdoor_max",
             "tol_pos_damper_oa",
             "flag_economizer_limit",
         ]
