+++ constrain/library/G36ReturnAirDamperPositionForReturnFanAirflowTracking.py
@@ -87,35 +87,30 @@ class G36ReturnAirDamperPositionForReturnFanAirflowTracking(RuleCheckBase):
         "output_coil_cooling",
         "position_damper_air_return",
         "position_damper_air_return_max",
-        "tol_pos_damper_ra",
         "position_damper_relief",
     ]
 
     def return_air_damper(self, data):
         if data["output_coil_heating"] > 0:
-            if (
-                abs(
-                    data["position_damper_air_return"]
-                    - data["position_damper_air_return_max"]
-                )
-                < data["tol_pos_damper_ra"]
-            ):
+            if abs(
+                data["position_damper_air_return"]
+                - data["position_damper_air_return_max"]
+            ) < self.get_tolerance("damper", "position"):
                 return True
             else:
                 return False
         elif data["output_coil_cooling"] > 0:
-            if data["position_damper_air_return"] < data["tol_pos_damper_ra"]:
+            if data["position_damper_air_return"] < self.get_tolerance(
+                "damper", "position"
+            ):
                 return True
             else:
                 return False
-        elif (
-            abs(
-                data["position_damper_air_return"]
-                - (1 - data["position_damper_relief"])
-                * data["position_damper_air_return_max"]
-            )
-            < data["tol_pos_damper_ra"]
-        ):
+        elif abs(
+            data["position_damper_air_return"]
+            - (1 - data["position_damper_relief"])
+            * data["position_damper_air_return_max"]
+        ) < self.get_tolerance("damper", "position"):
             return True
         else:
             return False
