+++ tests/test_g36_OutdoorAirDamperPositionForReliefDamperOrFan.py
@@ -7,36 +7,45 @@ import pandas as pd
 
 
 class TestG36OutdoorAirDamperPositionForReliefDamperOrFan(unittest.TestCase):
+    tolerances = {
+        "damper": {
+            "unit": "%",
+            "types": {
+                "position": 0.05,
+            },
+        }
+    }
+
     def test_return_air_damper_position(self):
         points = [
             "output_coil_heating",
             "output_coil_cooling",
             "position_damper_air_return",
             "position_damper_air_return_max",
-            "tol_pos_damper_ra",
             "position_damper_air_outdoor",
             "position_damper_air_outdoor_min",
             "position_damper_air_outdoor_max",
-            "tol_pos_damper_oa",
             "flag_economizer_limit",
         ]
         data = [
-            [1000, 0, 1.0, 1.0, 0.05, 0.8, 0.2, 1.0, 0.05, False],  # False
-            [1000, 0, 1.0, 1.0, 0.05, 0.2, 0.2, 1.0, 0.05, False],  # True
-            [0, 1000, 1.0, 1.0, 0.05, 0.2, 0.2, 1.0, 0.05, False],  # False
-            [0, 1000, 1.0, 1.0, 0.05, 0.8, 0.2, 1.0, 0.05, False],  # False
-            [0, 1000, 1.0, 1.0, 0.05, 0.98, 0.2, 1.0, 0.05, False],  # True
-            [0, 1000, 1.0, 1.0, 0.05, 0.8, 0.2, 1.0, 0.05, True],  # False
-            [0, 0, 1.0, 1.0, 0.05, 0.2, 0.2, 1.0, 0.05, False],  # False
-            [0, 0, 1.0, 1.0, 0.05, 0.8, 0.2, 1.0, 0.05, False],  # True
-            [0, 0, 0, 1.0, 0.05, 0.8, 0.2, 1.0, 0.05, False],  # False
-            [0, 0, 0, 1.0, 0.05, 1.0, 0.2, 1.0, 0.05, False],  # True
+            [1000, 0, 1.0, 1.0, 0.8, 0.2, 1.0, False],  # False
+            [1000, 0, 1.0, 1.0, 0.2, 0.2, 1.0, False],  # True
+            [0, 1000, 1.0, 1.0, 0.2, 0.2, 1.0, False],  # False
+            [0, 1000, 1.0, 1.0, 0.8, 0.2, 1.0, False],  # False
+            [0, 1000, 1.0, 1.0, 0.98, 0.2, 1.0, False],  # True
+            [0, 1000, 1.0, 1.0, 0.8, 0.2, 1.0, True],  # False
+            [0, 0, 1.0, 1.0, 0.2, 0.2, 1.0, False],  # False
+            [0, 0, 1.0, 1.0, 0.8, 0.2, 1.0, False],  # True
+            [0, 0, 0, 1.0, 0.8, 0.2, 1.0, False],  # False
+            [0, 0, 0, 1.0, 1.0, 0.2, 1.0, False],  # True
         ]
 
         df = pd.DataFrame(data, columns=points)
 
         verification_obj = run_test_verification_with_data(
-            "G36OutdoorAirDamperPositionForReliefDamperOrFan", df
+            "G36OutdoorAirDamperPositionForReliefDamperOrFan",
+            df,
+            tolerances=self.tolerances,
         )
 
         results = pd.Series(list(verification_obj.result))
