+++ tests/test_vav_minimum_turndown_during_reheat_pressure_reset.py
@@ -16,8 +16,6 @@ class TestVAVTurndown(unittest.TestCase):
             "flow_volumetric_air_max",
             "ratio_turndown_min",
             "pressure_duct_setpoint",
-            "tol_turndown",
-            "tol_p_press",
         ]
 
         timestamp = [
@@ -26,8 +24,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [False, 0.005, 0.01, 0.3, 1.0, 0.01, 0.01],
-            [False, 0.005, 0.01, 0.3, 1.1, 0.01, 0.01],
+            [False, 0.005, 0.01, 0.3, 1.0],
+            [False, 0.005, 0.01, 0.3, 1.1],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -51,8 +49,6 @@ class TestVAVTurndown(unittest.TestCase):
             "flow_volumetric_air_max",
             "ratio_turndown_min",
             "pressure_duct_setpoint",
-            "tol_turndown",
-            "tol_p_press",
         ]
 
         timestamp = [
@@ -61,8 +57,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [True, 0.005, 0.0, 0.3, 1.0, 0.01, 0.01],
-            [True, 0.005, 0.0, 0.3, 1.1, 0.01, 0.01],
+            [True, 0.005, 0.0, 0.3, 1.0],
+            [True, 0.005, 0.0, 0.3, 1.1],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -86,8 +82,6 @@ class TestVAVTurndown(unittest.TestCase):
             "flow_volumetric_air_max",
             "ratio_turndown_min",
             "pressure_duct_setpoint",
-            "tol_turndown",
-            "tol_p_press",
         ]
 
         timestamp = [
@@ -96,8 +90,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [True, 0.005, 0.0, 0.3, 1.0, 0.01, 0.01],
-            [True, 0.005, 0.0, 0.3, 1.0, 0.01, 0.01],
+            [True, 0.005, 0.0, 0.3, 1.0],
+            [True, 0.005, 0.0, 0.3, 1.0],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -121,8 +115,6 @@ class TestVAVTurndown(unittest.TestCase):
             "flow_volumetric_air_max",
             "ratio_turndown_min",
             "pressure_duct_setpoint",
-            "tol_turndown",
-            "tol_p_press",
         ]
 
         timestamp = [
@@ -131,8 +123,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
 
         data = [
-            [True, 0.005, 0.06, 0.3, 1.0, 0.01, 0.01],
-            [True, 0.005, 0.06, 0.3, 1.1, 0.01, 0.01],
+            [True, 0.005, 0.06, 0.3, 1.0],
+            [True, 0.005, 0.06, 0.3, 1.1],
         ]
 
         df = pd.DataFrame(data, columns=points, index=timestamp)
@@ -160,8 +152,6 @@ class TestVAVTurndown(unittest.TestCase):
             "flow_volumetric_air_max",
             "ratio_turndown_min",
             "pressure_duct_setpoint",
-            "tol_turndown",
-            "tol_p_press",
         ]
 
         timestamp = [
@@ -170,8 +160,8 @@ class TestVAVTurndown(unittest.TestCase):
         ]
[…]
