### 2025-03-28 d3125b0
add tolerance to freeze protection tests
