+++ tests/test_g36_FreezeProtectionStage3.py
@@ -7,8 +7,20 @@ from library import *
 import pandas as pd
 
 
-class TestG36FreezeProtectionStage2(unittest.TestCase):
-    def test_freeze_protection_2_pass(self):
+class TestG36FreezeProtectionStage3(unittest.TestCase):
+
+    tolerances = {
+        "damper": {
+            "unit": "%",
+            "types": {"position": 0.0, "command": 0.01, "general": 0.0},
+        },
+        "temperature": {
+            "unit": "deg. C",
+            "types": {"supply_air": 0.0, "general": 0.0},
+        },
+    }
+
+    def test_freeze_protection_3_pass(self):
         points = [
             "freeze_stat",
             "supply_air_temp",
@@ -43,7 +55,7 @@ class TestG36FreezeProtectionStage2(unittest.TestCase):
         expected_results = pd.Series([True, True, True, True, True, True])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage3", df
+            "G36FreezeProtectionStage3", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
@@ -51,7 +63,7 @@ class TestG36FreezeProtectionStage2(unittest.TestCase):
         self.assertTrue(results.equals(expected_results))
         self.assertTrue(binaryflag)
 
-    def test_freeze_protection_2_Fail(self):
+    def test_freeze_protection_3_Fail(self):
         points = [
             "freeze_stat",
             "supply_air_temp",
@@ -86,7 +98,7 @@ class TestG36FreezeProtectionStage2(unittest.TestCase):
         expected_results = pd.Series([True, False, False, False, False, False])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage3", df
+            "G36FreezeProtectionStage3", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
@@ -94,7 +106,7 @@ class TestG36FreezeProtectionStage2(unittest.TestCase):
         self.assertTrue(results.equals(expected_results))
         self.assertFalse(binaryflag)
 
-    def test_freeze_protection_2_pass(self):
+    def test_freeze_protection_3_untested(self):
         points = [
             "freeze_stat",
             "supply_air_temp",
@@ -129,7 +141,7 @@ class TestG36FreezeProtectionStage2(unittest.TestCase):
         expected_results = pd.Series([True, True, True, True, True, True])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage3", df
+            "G36FreezeProtectionStage3", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
