+++ tests/test_g36_FreezeProtectionStage1.py
@@ -9,6 +9,17 @@ import pandas as pd
 
 
 class TestG36FreezeProtectionStage1(unittest.TestCase):
+    tolerances = {
+        "damper": {
+            "unit": "%",
+            "types": {"position": 0.0, "command": 0.01, "general": 0.0},
+        },
+        "temperature": {
+            "unit": "deg. C",
+            "types": {"supply_air": 0.0, "general": 0.0},
+        },
+    }
+
     def test_freeze_protection_1_pass(self):
         points = [
             "temperature_air_supply_setpoint",
@@ -30,7 +41,7 @@ class TestG36FreezeProtectionStage1(unittest.TestCase):
         expected_results = pd.Series([True, True, True, True, True])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage1", df
+            "G36FreezeProtectionStage1", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
@@ -59,7 +70,7 @@ class TestG36FreezeProtectionStage1(unittest.TestCase):
         expected_results = pd.Series([True, False, False, False, True])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage1", df
+            "G36FreezeProtectionStage1", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
@@ -88,7 +99,7 @@ class TestG36FreezeProtectionStage1(unittest.TestCase):
         expected_results = pd.Series([True, True, True, True, True])
 
         verification_obj = run_test_verification_with_data(
-            "G36FreezeProtectionStage1", df
+            "G36FreezeProtectionStage1", df, tolerances=self.tolerances
         )
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
