+++ tests/test_g36_SimultaneousHeatingCooling.py
@@ -7,6 +7,15 @@ import pandas as pd
 
 
 class TestG36SimultaneousHeatingCooling(unittest.TestCase):
+    tolerances = {
+        "load": {
+            "unit": "W",
+            "types": {
+                "coil": 0.0,
+            },
+        },
+    }
+
     def test_heating_cooling(self):
         points = ["output_coil_heating", "output_coil_cooling"]
         data = [[1000, 0], [1000, 1000], [0, 1000]]
@@ -14,7 +23,7 @@ class TestG36SimultaneousHeatingCooling(unittest.TestCase):
         df = pd.DataFrame(data, columns=points)
 
         verification_obj = run_test_verification_with_data(
-            "G36SimultaneousHeatingCooling", df
+            "G36SimultaneousHeatingCooling", df, tolerances=self.tolerances
         )
 
         results = pd.Series(list(verification_obj.result))
