+++ constrain/library/LocalLoopSaturationDirectActingMax.py
@@ -75,7 +75,11 @@ class LocalLoopSaturationDirectActingMax(RuleCheckBase):
     points = ["value_sensor", "value_setpoint", "command_control", "command_max"]
 
     def saturation_flag(self, t):
-        if 0 <= t["command_max"] - t["command_control"] <= 0.01:
+        if (
+            0
+            <= t["command_max"] - t["command_control"]
+            <= self.get_tolerance("command", "general")
+        ):
             return True
         else:
             return False
