+++ tests/test_vav_turndown_during_reheat.py
@@ -9,6 +9,15 @@ from lib_unit_test_runner import *
 
 
 class TestVAVTurndown(unittest.TestCase):
+    tolerances = {
+        "ratio": {
+            "unit": "%",
+            "types": {
+                "flow": 0.01,
+            },
+        }
+    }
+
     def test_vav_turndown_during_reheat_pass(self):
         points = [
             "flag_coil_reheat",
@@ -33,7 +42,7 @@ class TestVAVTurndown(unittest.TestCase):
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
         verification_obj = run_test_verification_with_data(
-            "VAVTurndownDuringReheat", df
+            "VAVTurndownDuringReheat", df, tolerances=self.tolerances
         )
 
         results = pd.Series(list(verification_obj.result))
@@ -74,7 +83,7 @@ class TestVAVTurndown(unittest.TestCase):
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
         verification_obj = run_test_verification_with_data(
-            "VAVTurndownDuringReheat", df
+            "VAVTurndownDuringReheat", df, tolerances=self.tolerances
         )
 
         results = pd.Series(list(verification_obj.result))
@@ -115,7 +124,7 @@ class TestVAVTurndown(unittest.TestCase):
         df = pd.DataFrame(data, columns=points, index=timestamp)
 
         verification_obj = run_test_verification_with_data(
-            "VAVTurndownDuringReheat", df
+            "VAVTurndownDuringReheat", df, tolerances=self.tolerances
         )
         results = list(verification_obj.result)
         expected_results = [
