### 2025-03-28 c939ef9
Add new action to test build and installation of ConStrain.
