+++ .github/workflows/test_deployment.yml
@@ -1,0 +1,30 @@
+name: Test installation of ConStrain on different OS
+on: workflow_dispatch
+jobs:
+    main:
+        name: Build package with Poetry and run example
+        runs-on: ['ubuntu-latest', 'windows-latest', 'macos-latest']
+        steps:
+            - uses: actions/checkout@v3
+            - name: Install Python
+              uses: actions/setup-python@v4
+            - name: Install poetry
+              uses: abatilo/actions-poetry@v2
+            - name: Setup a local virtual environment (if no poetry.toml file)
+              run: |
+                poetry config virtualenvs.create true --local
+                poetry config virtualenvs.in-project true --local
+            - uses: actions/cache@v3
+              name: Define a cache for the virtual environment based on the dependencies lock file
+              with:
+                path: ./.venv
+                key: venv-${{ hashFiles('poetry.lock') }}
+            - name: Install the project dependencies
+              run: poetry install
+            - name: Build package
+              run: poetry build
+            - name: Install package
+              run: python -m pip install --find-links=.\dist\ constrain
+            - name: Run sample ConStrain scrip
+              run: python -m .\sample_constrain_run.py
+            
