+++ .github/workflows/test_deployment.yml
@@ -24,7 +24,7 @@ jobs:
             - name: Build package
               run: poetry build
             - name: Install package
-              run: pip install .dist constrain
+              run: pip install --find-links=.\dist\ constrain
             - name: Run sample ConStrain scrip
-              run: python -m .\sample_constrain_run.py
+              run: python -m .github\workflows\sample_constrain_run.py
             
