### 2025-03-28 134a54e
Try alternative installation.
