+++ .github/workflows/test_deployment.yml
@@ -24,7 +24,8 @@ jobs:
             - name: Build package
               run: poetry build
             - name: Install package
-              run: find ./dist/*.whl | xargs pip install
+              #run: find ./dist/*.whl | xargs pip install
+              run: python -c "import glob, subprocess; [subprocess.call(['pip', 'install', f]) for f in glob.glob('./dist/*.whl')]"
             - name: Run sample ConStrain scrip
               run: python .github/workflows/sample_constrain_run.py
             
