+++ .github/workflows/test_deployment.yml
@@ -3,7 +3,7 @@ on: workflow_dispatch
 jobs:
     main:
         name: Build package with Poetry and run example Constrin script
-        runs-on: ['windows-latest']
+        runs-on: ['macos-latest']
         steps:
             - uses: actions/checkout@v3
             - name: Install Python
