+++ pyproject.toml
@@ -14,7 +14,7 @@ matplotlib = "^3.8.2"
 seaborn = "^0.13.0"
 pandas = "^2.1.3"
 eppy = "^0.5.63"
-fuzzywuzzy = "^0.18.0"
+#fuzzywuzzy = "^0.18.0"
 scikit-learn = "^1.3.2"
 scipy = "^1.11.4"
 tqdm = "^4.66.1"
