+++ .github/workflows/test_deployment.yml
@@ -6,11 +6,14 @@ jobs:
         runs-on: ${{ matrix.os }}
         strategy:
             matrix:
-                os: [ubuntu-latest, windows-latest]
+                os: [ubuntu-latest, windows-latest, macos-latest]
+                python-version: ["3.10"]
         steps:
             - uses: actions/checkout@v3
-            - name: Install Python
+            - name: Set up Python ${{ matrix.python-version }}
               uses: actions/setup-python@v4
+              with:
+                python-version: ${{ matrix.python-version }}
             - name: Install poetry
               uses: abatilo/actions-poetry@v2
             - name: Setup a local virtual environment (if no poetry.toml file)
