+++ .github/workflows/test_deployment.yml
@@ -32,5 +32,5 @@ jobs:
             - name: Install package
               run: python -c "import glob, subprocess; [subprocess.call(['pip', 'install', f]) for f in glob.glob('./dist/*.whl')]"
             - name: Run sample ConStrain script
-              run: cd ..; python ./constrain/.github/workflows/sample_constrain_run.py
+              run: cd ..; python ./ConStrain/.github/workflows/sample_constrain_run.py
             
