### 2025-03-28 c89573e
Clarification for multi-threading.
