### 2025-03-29 e0b10a3
Fixed number_occupants < 0 in pseudo code
