+++ .github/workflows/test_deployment.yml
@@ -6,7 +6,7 @@ jobs:
         runs-on: ${{ matrix.os }}
         strategy:
             matrix:
-                os: [ubuntu-latest, windows-latest, macos-latest]
+                os: [ubuntu-latest] #, windows-latest, macos-latest]
                 python-version: ["3.10"]
         steps:
             - uses: actions/checkout@v3
