+++ .github/workflows/unit_tests.yml
@@ -1,7 +1,7 @@
+name: Unit Tests
 on:
   push:
     branches: [ '*' ]
-
 jobs:
   black-formatting-check:
     name: Black formatting check
