+++ constrain/library/G36CoolingOnlyTerminalBoxDeadbandAirflowSetpoint.py
@@ -41,15 +41,15 @@ end
 
 ### Data requirements
 
-- mode_system: System operation mode
+- mode_system: System operation mode (occupied, cooldown, setup, warmup, setback, unoccupied)
   - Data Value Unit: enumeration
   - Data Point Affiliation: System control
 
-- state_zone: Zone state
+- state_zone: Zone state (if state_zone is not "deadband", this verification item falls into the "untested" result)
   - Data Value Unit: enumeration
   - Data Point Affiliation: Zone control
 
-- flow_volumetric_air_setpoint_min: Minimum airflow
+- flow_volumetric_air_setpoint_min: Minimum airflow setpoint
   - Data Value Unit: volumetric flow rate
   - Data Point Affiliation: Zone airflow control
 
