+++ constrain/library/G36ReheatTerminalBoxHeatingCoilTracking.py
@@ -49,7 +49,7 @@ else:
 
 ### Data requirements
 
-- mode_system: System operation mode
+- mode_system: System operation mode (if mode_system is not "heating", this verification item falls into the "untested" result)
   - Data Value Unit: enumeration
   - Data Point Affiliation: System control
 
