### 2025-03-30 5659d80
Fixed G36FreezeProtectionStage3 docstring
