### 2025-03-30 f04e166
Fixed G36FreezeProtectionStage3 pseudocode
