### 2025-03-30 accce9e
Fixed - position_damper description in docstring
