+++ constrain/library/G36MinOAwoEconomizer.py
@@ -60,11 +60,11 @@ else:
   - Data Value Unit: °C
   - Data Point Affiliation: Economizer control
 
-- position_damper_air_outdoor: Outdoor air damper command
+- position_damper_air_outdoor: Outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_return: Return air damper command
+- position_damper_air_return: Return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
