+++ constrain/library/G36OutdoorAirDamperPositionForReliefDamperOrFan.py
@@ -70,23 +70,23 @@ else:
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_return: Return air damper command
+- position_damper_air_return: Return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_return_max: Maximum return air damper command
+- position_damper_air_return_max: Maximum return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_outdoor: Outdoor air damper command
+- position_damper_air_outdoor: Outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_outdoor_min: Minimum outdoor air damper command
+- position_damper_air_outdoor_min: Minimum outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_outdoor_max: Maximum outdoor air damper command
+- position_damper_air_outdoor_max: Maximum outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
