+++ constrain/library/G36ReturnAirDamperPositionForReturnFanDirectBuildingPressure.py
@@ -55,11 +55,11 @@ else:
   - Data Value Unit: percent
   - Data Point Affiliation: System control
 
-- position_damper_air_return: Return air damper command
+- position_damper_air_return: Return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_return_max: Maximum return air damper command
+- position_damper_air_return_max: Maximum return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
