+++ constrain/library/G36FreezeProtectionStage1.py
@@ -43,11 +43,11 @@ if never (temperature_air_supply_setpoint < 4.4 (continuously 5 minutes)):
   - Data Value Unit: °C
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_outdoor: Outdoor air damper command
+- position_damper_air_outdoor: Outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_outdoor_min: Minimum outdoor air damper command
+- position_damper_air_outdoor_min: Minimum outdoor air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
