+++ constrain/library/G36ReliefAirDamperPositionForReturnFanAirflowTracking.py
@@ -55,15 +55,15 @@ else:
   - Data Value Unit: percent
   - Data Point Affiliation: System control
 
-- position_damper_relief: Relief air damper command
+- position_damper_relief: Relief air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_relief_max: Maximum relief air damper command
+- position_damper_relief_max: Maximum relief air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
-- position_damper_air_return: Return air damper command
+- position_damper_air_return: Return air damper position
   - Data Value Unit: percent
   - Data Point Affiliation: Air handling unit
 
