### 2025-03-30 bbd73eb
Fixed the old var name in library.json file & Fixed abs logic in pseudo code
