+++ constrain/library/G36ReturnAirDamperPositionForReturnFanDirectBuildingPressure.py
@@ -30,12 +30,12 @@ The verification checks return air damper position under three conditions:
 
 ```python
 if output_coil_heating > 0:
-    if abs(position_damper_air_return - position_damper_air_return_max) < 0:
+    if abs(position_damper_air_return - position_damper_air_return_max) = 0:
         pass
     else:
         fail
 elif output_coil_cooling > 0:
-    if abs(position_damper_air_return - 0) < 0:
+    if abs(position_damper_air_return - 0) = 0:
         pass
     else:
         fail
