+++ constrain/library/G36ReheatTerminalBoxDeadbandAirflowSetpoint.py
@@ -37,7 +37,7 @@ else:
         case 'cooldown', 'setup', 'warmup', 'setback', 'unoccupied':
             minimum = 0
 
-    if abs(flow_volumetric_air_setpoint - minimum) <= 0:
+    if abs(flow_volumetric_air_setpoint - minimum) = 0:
         pass
     else:
         fail
