### 2025-03-30 aaff51a
Fixed zone temp control pseudo code
