### 2025-03-30 cb29d33
fix chdir logic
