+++ constrain/library/vav_minimum_turndown_during_reheat_pressure_reset.py
@@ -54,7 +54,7 @@ else:
 ```
 
 ### Data requirements
-- flag_coil_reheat: VAV box reheat coil operation status
+- flag_coil_reheat: VAV box reheat coil operation flag
   - Data Value Unit: binary
   - Data Point Affiliation: Terminal unit control
 
