+++ constrain/library/MZSystemOccupiedStandbyVentilationZoneControl.py
@@ -46,7 +46,7 @@ class MZSystemOccupiedStandbyVentilationZoneControl(RuleCheckBase):
     ]
     last_non_standby_mode_requested_m_oa = None  # expects kg/s
 
-    def occupied_standby_ventilation_zontrol_control(self, data):
+    def occupied_standby_ventilation_zone_control(self, data):
         # initialization
         if self.last_non_standby_mode_requested_m_oa is None:
             self.last_non_standby_mode_requested_m_oa = data["m_oa_requested_by_system"]
@@ -68,5 +68,5 @@ class MZSystemOccupiedStandbyVentilationZoneControl(RuleCheckBase):
 
     def verify(self):
         self.result = self.df.apply(
-            lambda d: self.occupied_standby_ventilation_zontrol_control(d), axis=1
+            lambda d: self.occupied_standby_ventilation_zone_control(d), axis=1
         )
