+++ tools/wizard/steps/verification_case_selection.py
@@ -3,7 +3,7 @@ import json
 from PyQt6 import QtWidgets
 
 from tools.wizard.utils.load_schemas import load_verification_cases_library, load_verification_cases_schema
-
+from tools.wizard.steps import WizardPageIds
 class VerificationCaseSelectionPage(QtWidgets.QWizardPage):
     def __init__(self, parent=None):
         super().__init__(parent)
@@ -19,9 +19,12 @@ class VerificationCaseSelectionPage(QtWidgets.QWizardPage):
         
         layout = QtWidgets.QVBoxLayout()
         
+        self.searchBox = QtWidgets.QLineEdit()
+        self.searchBox.setPlaceholderText("Search")
+        self.searchBox.textChanged.connect(self.filterTree)
+
         self.tree = QtWidgets.QTreeWidget()
         self.tree.setHeaderLabels(["Available Items"])
-
         for key in self._verification_cases_library.keys():
             item = QtWidgets.QTreeWidgetItem([key])
             self.tree.addTopLevelItem(item)
@@ -30,13 +33,27 @@ class VerificationCaseSelectionPage(QtWidgets.QWizardPage):
         
         self.selectionLabel = QtWidgets.QLabel("Please select an item from the list.")
 
+        layout.addWidget(self.searchBox)
         layout.addWidget(self.tree)
         layout.addWidget(self.selectionLabel)
         
         self.setLayout(layout)
 
+    def filterTree(self):
+        filter_text = self.searchBox.text().lower()
+        for i in range(self.tree.topLevelItemCount()):
+            item = self.tree.topLevelItem(i)
+            item.setHidden(filter_text not in item.text(0).lower())
+
     def onItemSelected(self, current, previous):
         if current:
             selected_item = current.text(0)
             self.selectionLabel.setText(f"Selected Item: {selected_item}")
-            self.wizard().selected_verification_class = selected_item
+
+            if current != previous:
+                self.wizard().selected_verification_class = selected_item
+                variable_mapping_page = self.wizard().page(WizardPageIds.VARIABLE_MAPPING.value)
+                variable_mapping_page.reset_mappings()
+
+    def nextId(self):
+        return WizardPageIds.CSV_UPLOAD.value
