### 2025-04-01 c483136
Update resource load method for executable
