+++ tools/wizard/utils/load_schemas.py
@@ -1,14 +1,24 @@
 import json
+import os
+import sys
 
-VERIFICATION_CASES_SCHEMA_PATH = "schema/verification_cases.schema.json"
-LIBRARY_PATH = "schema/library.json"
 
-def load_verification_cases_schema():
-    return load_json(VERIFICATION_CASES_SCHEMA_PATH)
-    
+def get_resource_path(relative_path):
+    if getattr(sys, "frozen", False):
+        base_path = sys._MEIPASS
+    else:
+        base_path = os.path.abspath(".")
+
+    return os.path.join(base_path, relative_path)
+
+
+LIBRARY_PATH = get_resource_path("schema/library.json")
+
+
 def load_verification_cases_library():
     return load_json(LIBRARY_PATH)
-    
+
+
 def load_json(file_path):
     with open(file_path, "r") as file:
         return json.load(file)
