+++ tools/wizard/components/verification_case_runner.py
@@ -6,7 +6,7 @@ from constrain.app.submit import EmittingStream
 
 class VerificationCaseRunner(QtCore.QThread):
     update_text = QtCore.pyqtSignal(str)
-    
+
     def __init__(self, verification: Verification):
         super(VerificationCaseRunner, self).__init__()
         self.verification = verification
@@ -17,4 +17,3 @@ class VerificationCaseRunner(QtCore.QThread):
             self.verification.run()
         finally:
             sys.stdout = sys.__stdout__  # Reset stdout to original
-            
