+++ tools/wizard/steps/csv_upload.py
@@ -2,6 +2,8 @@ from PyQt6 import QtWidgets
 import pandas as pd
 
 from tools.wizard.steps import WizardPageIds
+
+
 class CSVUploadPage(QtWidgets.QWizardPage):
     def __init__(self, parent=None):
         super().__init__(parent)
@@ -20,12 +22,14 @@ class CSVUploadPage(QtWidgets.QWizardPage):
 
         layout.addWidget(self.uploadButton)
         layout.addWidget(self.filePathLabel)
-        
+
         self.setLayout(layout)
 
     def uploadCSV(self):
         options = QtWidgets.QFileDialog.Option.ReadOnly
-        fileName, _ = QtWidgets.QFileDialog.getOpenFileName(self, "Open CSV", "", "CSV Files (*.csv);;All Files (*)", options=options)
+        fileName, _ = QtWidgets.QFileDialog.getOpenFileName(
+            self, "Open CSV", "", "CSV Files (*.csv);;All Files (*)", options=options
+        )
         if fileName:
             self.uploaded_file_path = fileName
             self.filePathLabel.setText(f"Uploaded: {fileName}")
@@ -33,12 +37,14 @@ class CSVUploadPage(QtWidgets.QWizardPage):
             try:
                 self.wizard().data = pd.read_csv(self.uploaded_file_path)
             except Exception as e:
-                QtWidgets.QMessageBox.critical(self, "Error", f"Failed to read CSV: {str(e)}")
+                QtWidgets.QMessageBox.critical(
+                    self, "Error", f"Failed to read CSV: {str(e)}"
+                )
                 return
             self.completeChanged.emit()
-        
+
     def isComplete(self):
         return self.uploaded_file_path is not None
-    
+
     def nextId(self):
         return WizardPageIds.VARIABLE_MAPPING.value
