+++ constrain/checklib.py
@@ -131,7 +131,8 @@ class CheckLibBase(ABC):
         return {
             "md_content": md_content,
             "outcome_notes": outcome_dict,
-            "model_file": item_dict["simulation_IO"].get("idf", "")
+            "model_file": item_dict["simulation_IO"]
+            .get("idf", "")
             .split("/")[-1]
             .split("\\")[-1]
             .replace(".idf", ""),
