### 2025-04-01 e922897
Remove verification cases schema from build
