### 2025-04-04 4a30543
Add restart option
