+++ tools/wizard/components/verification_case_runner.py
@@ -1,7 +1,8 @@
 from PyQt6 import QtCore
-import sys
+import sys, os
 from constrain.api.verification import Verification
 from constrain.app.submit import EmittingStream
+from constrain.api.reporting import Reporting
 
 
 class VerificationCaseRunner(QtCore.QThread):
@@ -15,6 +16,12 @@ class VerificationCaseRunner(QtCore.QThread):
         sys.stdout = EmittingStream(self.update_text)
         try:
             self.verification.run()
+            reporting = Reporting(
+                verification_json=f"{os.getcwd()}/*_md.json",
+                result_md_name="report_summary.md",
+                report_format="markdown",
+            )
+            reporting.report_multiple_cases()
         finally:
             print("Verification Complete")
             sys.stdout = sys.__stdout__  # Reset stdout to original
