+++ tools/wizard/steps/run_verification.py
@@ -77,12 +77,6 @@ class RunVerificationPage(QtWidgets.QWizardPage):
             fig_size=(6, 5),
             num_threads=2,
         )
-        reporting = Reporting(
-            verification_json="./*_md.json",
-            result_md_name="report_summary.md",
-            report_format="markdown",
-        )
-        reporting.report_multiple_cases()
         self.verification_runner = VerificationCaseRunner(verification)
         self.verification_runner.update_text.connect(self.append_output)
         self.verification_runner.finished.connect(self.cleanup_tmpdir)
