### 2025-04-20 #7195 @lazlop
> 👋 [@dhblum](https://github.com/dhblum) & [@lazlop](https://github.com/lazlop), would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I am able to review this, what are next steps?
