### 2025-04-20 #8083 @editorialbot
[REVIEW]: Control Strainer (ConStrain): a data-driven control verification framework
**Submitting author:** @leijerry888 (<a href="http://orcid.org/0000-0003-3310-9750">Xuechen Lei</a>)
**Repository:** https://github.com/pnnl/ConStrain
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.4.0
