### 2025-04-22 9a2b2f9
Misc. typo in docstring.
