### 2025-04-30 c3226a3
undo changes to /test/api/test_data_adapter.py
