### 2025-04-30 5cd1582
Added Modelica dataset unit test and use global tolerance
