+++ tests/test_trim_respond_logic.py
@@ -3,8 +3,9 @@ import unittest
 
 import pandas as pd
 
-sys.path.append("./src")
-from library.trim_respond_logic import TrimRespondLogic
+sys.path.append("./constrain")
+from lib_unit_test_runner import *
+from library import *
 
 # data preparation
 start_date = "2023-06-21 12:00:00"  # Start date and time
@@ -16,7 +17,14 @@ data = pd.DataFrame(
 )
 # fmt: off
 # The below data is from Figure 5.1.14.4 in the ASHRAE G36-2021
-data["setpoint"] = [0.50, 0.46, 0.42, 0.48,0.60, 0.75, 0.81, 0.77, 0.73, 0.69, 0.65, 0.61, 0.57, 0.53, 0.49, 0.45, 0.41, 0.37, 0.33, 0.29, 0.25, 0.21, 0.36, 0.51, 0.66, 0.81, 0.77, 0.73, 0.85, 0.81]
+values_in_inches = [0.50, 0.46, 0.42, 0.48,0.60, 0.75, 0.81, 0.77, 0.73, 0.69, 0.65, 0.61, 0.57, 0.53, 0.49, 0.45, 0.41, 0.37, 0.33, 0.29, 0.25, 0.21, 0.36, 0.51, 0.66, 0.81, 0.77, 0.73, 0.85, 0.81]
+
+# Define the conversion factor
+in_to_pa = 248.84
+
+# Convert to Pascals
+data["setpoint"]  = [value * in_to_pa for value in values_in_inches]
+
 data["number_of_requests"] = [0, 1, 2, 3, 4, 6, 3, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 1, 0, 1, 2, 6, 6, 5, 5, 2, 2, 4, 2]
 # fmt: on
 
@@ -29,16 +37,17 @@ class TestTrimRespond(unittest.TestCase):
         with self.assertLogs() as logobs:
             TrimRespondLogic(
                 dict(data),
-                Td="0",
+                Td=0,
                 ignored_requests=2,
-                SP0=0.5,
-                SPtrim=-0.04,
-                SPres=0.06,
-                SPmin=0.15,
-                SPmax=1.5,
-                SPres_max=0.15,
-                tol=0.01,
+                SP0=120,
+                SPtrim=-10,
+                SPres=15,
+                SPmin=37,
+                SPmax=370,
+                SPres_max=37,
                 controller_type="direct_acting",
+                variable_type="pressure",
+                variable_subtype="static",
             )
             self.assertEqual(
                 "ERROR:root:The type of the `df` arg must be a dataframe. It cannot be <class 'dict'>.",
@@ -57,8 +66,9 @@ class TestTrimRespond(unittest.TestCase):
                 SPmin=0.15,
                 SPmax=1.5,
                 SPres_max=0.15,
-                tol=0.01,
                 controller_type="direct_acting",
+                variable_type="pressure",
+                variable_subtype="static",
             )
             self.assertEqual(
                 "ERROR:root:Index's format is not in datetime format.",
@@ -69,16 +79,17 @@ class TestTrimRespond(unittest.TestCase):
         with self.assertLogs() as logobs:
             TrimRespondLogic(
                 data.drop("setpoint", axis=1),
-                Td="0",
+                Td=0,
                 ignored_requests=2,
-                SP0=0.5,
-                SPtrim=-0.04,
-                SPres=0.06,
-                SPmin=0.15,
-                SPmax=1.5,
-                SPres_max=0.15,
-                tol=0.01,
+                SP0=120,
+                SPtrim=-10,
+                SPres=15,
+                SPmin=37,
+                SPmax=370,
+                SPres_max=37,
                 controller_type="direct_acting",
+                variable_type="pressure",
[…]
