### 2025-05-02 541b15b
Replaced `mode_system` with `mode_operation`
