+++ tests/test_g36_CoolingOnlyTerminalBoxCoolingAirflowSetpoint.py
@@ -11,7 +11,7 @@ import numpy as np
 class TestG36CoolingOnlyTerminalBoxCoolingAirflowSetpoint(unittest.TestCase):
     def test_g36_cooling_only_terminal_box_cooling_airflow_setpoint(self):
         points = [
-            "mode_system",
+            "mode_operation",
             "state_zone",
             "flow_volumetric_air_cool_max",
             "flow_volumetric_air_setpoint_min",
