+++ tests/test_g36_ReheatTerminalBoxHeatingCoilLowerBound.py
@@ -11,7 +11,7 @@ import numpy as np
 class TestG36ReheatTerminalBoxHeatingCoilLowerBound(unittest.TestCase):
     def test_g36_reheat_terminal_box_heating_coil_lower_bound(self):
         points = [
-            "mode_system",
+            "mode_operation",
             "command_coil_heat",
             "temperature_air_discharge",
         ]
