+++ constrain/library/trim_respond_logic.py
@@ -39,7 +39,9 @@ def TrimRespondLogic(
     """Trim and respond logic verification & setpoint calculation.
 
     Args:
-        df (dataframe): dataframe that must include timestamp (`Date/Time` column), time-series data of setpoint (`setpoint` column), number of requests (`number_of_requests` column name),
+        df (dataframe): dataframe that must include timestamp (`Date/Time` column),
+                        time-series data of setpoint (`setpoint` column), number of requests (`number_of_requests` column name),
+                        flag_device (int or bool): flag to indicate device status (e.g., AHU). This can be integer (0: off, 1: on) or boolean (True, False),
         Td (float or int): Time delay in minutes.
         ignored_requests (int): Number of ignored requests.
         SP0 (float or int): Initial setpoint.
@@ -56,7 +58,8 @@ def TrimRespondLogic(
                                     airflow: supply, outdoor_air, exhaust_air, general
                                     waterflow: hot, general
                                     pressure: static, building, general
-    Return: dataframe including verification in boolean and setpoint calculation results in each timestep.
+
+    Return: dataframe including verification in boolean and calculated setpoint calculation results in each timestep.
     """
 
     # check df type
@@ -72,7 +75,7 @@ def TrimRespondLogic(
         return None
 
     # check df columns
-    for col in ("setpoint", "number_of_requests"):
+    for col in ("setpoint", "number_of_requests", "flag_device"):
         if col not in df.columns:
             logging.error(f"{col} column doesn't exist in the `df`.")
             return None
@@ -101,6 +104,7 @@ def TrimRespondLogic(
             f"The type of the `SPtrim` arg must be a float or int. It cannot be {type(SPtrim)}."
         )
         return None
+
     if not isinstance(SPres, (float, int)):
         logging.error(
             f"The type of the `SPres` arg must be a float or int. It cannot be {type(SPres)}."
@@ -132,7 +136,7 @@ def TrimRespondLogic(
         )
         return None
 
-    # set tolerance
+    # Check tolerance input arguments
     if variable_type not in VARIABLE_SET:
         logging.error(
             f"The `variable_type` arg must be one of temperature, airflow, waterflow, pressure. It can't be `{variable_type}`."
@@ -145,115 +149,139 @@ def TrimRespondLogic(
         )
         return None
 
+    # Read the tolerance JSON file
     path_to_custom_tolerance_file = Path(__file__).parent.parent / "tolerances.json"
     with open(path_to_custom_tolerance_file) as f:
         tolerances = json.load(f)
 
+    # Define tolerance
     tol = tolerances[variable_type]["types"][variable_subtype]
-
-    # calculate actual start time
-    initial_timestamp = df.index[0]
-    actual_start_time = initial_timestamp + pd.Timedelta(minutes=Td)
 
     # create "result" dataframe
     result = pd.DataFrame(
-        columns=["verification", "setpoint"],
-        index=df[df.index >= actual_start_time].index,
+        columns=["verification", "calculated_setpoint"],
+        index=df.index,
     )
-
-    # preprocess df, cut off before Td
-    copied_df = df.copy()[df.index > actual_start_time]
-
-    # setup an initial value
-    result.loc[actual_start_time, "verification"] = True
-    if initial_timestamp == actual_start_time:
-        result.loc[actual_start_time, "setpoint"] = SP0
-    else:
-        try:
-            df.loc[actual_start_time, "setpoint"]
-        except KeyError:
-            logging.error(
-                f"The delayed timestamp must be included in the `df` timestamp."
-            )
-            return None
 
     # start the T&R logic verification
     SPtrim = abs(SPtrim)
[…]
