+++ tests/test_trim_respond_logic.py
@@ -31,7 +31,7 @@ data["flag_device"] = [1 for _ in range(30)]
 
 class TestTrimRespond(unittest.TestCase):
     def test_check_args_type(self):
-        """Test whether arguments' type is correct."""
+        """Test if argument types are correct."""
 
         # check `data` (type)
         with self.assertLogs() as logobs:
@@ -328,7 +328,7 @@ class TestTrimRespond(unittest.TestCase):
             )
 
     def test_TR_logic_verification(self):
-        """test whether the T&R logic was implemented correctly."""
+        """Test if the T&R logic was implemented correctly."""
 
         # verify the verification was implemented correctly
         tr_obj = TrimRespondLogic(
@@ -349,54 +349,31 @@ class TestTrimRespond(unittest.TestCase):
         # check if all verification passed
         self.assertTrue(all(tr_obj["verification"]))
 
-    def test_TR_winter_day_reset(self):
-        """Test winter day reset dataset from Modelica"""
+    def test_TR_winter_day_reset_singlesample(self):
+        """Test the winter day reset dataset from Modelica"""
 
         start_date = "2023-12-21 12:00:00"
         modelica_winter = pd.DataFrame(
-            columns=["setpoint", "number_of_requests"],
+            columns=["setpoint", "number_of_requests", "flag_device"],
             index=pd.date_range(start=start_date, periods=1441, freq="min"),
         )
-
+        # fmt: off
         modelica_winter["setpoint"] = (
-            [120] * 295
-            + [108] * 3
-            + [96] * 2
-            + [99] * 2
-            + [119] * 2
-            + [139] * 2
-            + [159] * 2
-            + [179] * 2
-            + [199] * 2
-            + [219] * 2
-            + [222] * 2
-            + [210] * 2
-            + [198] * 2
-            + [186] * 2
-            + [174] * 2
-            + [162] * 2
-            + [150] * 2
-            + [138] * 2
-            + [126] * 2
-            + [114] * 2
-            + [102] * 2
-            + [90] * 2
-            + [78] * 2
-            + [66] * 2
-            + [54] * 2
-            + [42] * 2
-            + [30] * 2
-            + [25] * 792
+            [120] * 294
+            + [
+                108, 96, 84, 72, 60, 48, 51, 71, 91, 111, 131, 151, 171, 191, 211, 231,
+                251, 271, 291, 311,  314, 317, 305, 293, 281, 269, 257, 245, 233, 221,
+                209, 197, 185, 173, 161, 149, 137, 125, 113, 101, 89, 29, 77, 65, 53, 41,
+            ]
+            + [25] * 800
             + [120] * 301
         )
+        # fmt: on
         modelica_winter["number_of_requests"] = (
             [0] * 300
             + [3] * 1
-            + [9] * 9
-            + [8] * 2
-            + [6] * 2
-            + [3] * 2
-            + [1] * 2
+            + [9] * 3
+            + [10, 12, 11, 11, 11, 11, 10, 10, 7, 7, 3, 3, 2, 2]
             + [0] * 1123
         )
         modelica_winter["flag_device"] = [0] * 283 + [1] * 857 + [0] * 301
@@ -417,51 +394,8 @@ class TestTrimRespond(unittest.TestCase):
             variable_subtype="static",
         )
 
-        # check if all verification passed
-        self.assertTrue(all(tr_obj["verification"]))
-
-    def test_TR_summer_day_reset(self):
-        """Test summer day reset dataset from Modelica"""
-
-        start_date = "2023-06-21 12:00:00"
-        modelica_summer = pd.DataFrame(
-            columns=["setpoint", "number_of_requests", "flag_device"],
-            index=pd.date_range(start=start_date, periods=1441, freq="min"),
-        )
-
[…]
