### 2025-05-21 #8083 @AdamRJensen
> @AdamRJensen - The license they have is not a typical OSI license, I would like to point this out and ask if this is their final  license. Should I do this in in issue on the ConStrain repository?

That seems like a good option. Thanks for pointing this out.
