### 2025-05-21 pnnl/ConStrain#87 @lazlop
JOSS Paper Issues
- OSS license is not one through OSI. It seems to have permissions lining up with other open source licenses. Is this the final license for this software, or will something more traditional (e.g. BSD, MIT) be used?
- Based on quick-start guide, the software appears to function as expected.
- Example usage documentation is insufficient
    - GUI is not documented
    - The readme mentions demo/api_demo, but I do not see a folder demo in the linked repository (the develop branch)
    - The plotting generated when following “Using ConStrain as a python library” is not described, and I am not easily able to interpret them.
    - I can’t verify all functionality
    - Local Loop Verification Items List and Brick integration doc in the readme link to missing files - I am not sure how to test the functionality for these things.
    - I think there needs to be a tutorial perhaps from a building operator perspective so I can understand how each of the elements of functionality are intended to be used, not just what is in the quickstart guide.
    - How simulation is supposed to be used is not described. Are users of the software, those mentioned in the statement of need, intended to do something with simulation? Is this just so developers can test the software or add new verifications?
- The quick start guide is quite good, but the example usage is not relevant enough to real-world problems. There needs to be a set of tutorials targetted at some of the folks that would be interested in the framework, at least Cx agents.
- In the white-paper, I think there needs to be more discussion comparing this tool to what the state of the industry is. How does this improve upon what tools Cx agents may already be using? Is this the only open source software for controls verification?
