### 2025-05-27 #8083 @AdamRJensen
> ```
> @editorialbot generate my checklist
> ```
@FWuellhorst the comment you made has the editorialbot command enclosed in backticks which is the cause of the issue. If you post the comment again without the backticks, I am confident it should work! 😄
