+++ tests/test_trim_respond_logic.py
@@ -1,6 +1,7 @@
 import sys
 import unittest
 
+import numpy as np
 import pandas as pd
 
 sys.path.append("./constrain")
@@ -8,25 +9,85 @@ from lib_unit_test_runner import *
 from library import *
 
 # data preparation
-start_date = "2023-06-21 12:00:00"  # Start date and time
-end_date = "2023-06-21 12:59:59"  # End date and time
+start_date = "2023-06-21 12:00:00"
+end_date = "2023-06-21 12:59:59"
 
 data = pd.DataFrame(
     columns=["setpoint", "number_of_requests"],
     index=pd.date_range(start=start_date, end=end_date, freq="2min"),
 )
-# fmt: off
 # The below data is from Figure 5.1.14.4 in the ASHRAE G36-2021
-values_in_inches = [0.50, 0.46, 0.42, 0.48,0.60, 0.75, 0.81, 0.77, 0.73, 0.69, 0.65, 0.61, 0.57, 0.53, 0.49, 0.45, 0.41, 0.37, 0.33, 0.29, 0.25, 0.21, 0.36, 0.51, 0.66, 0.81, 0.77, 0.73, 0.85, 0.81]
+values_in_inches = [
+    0.50,
+    0.46,
+    0.42,
+    0.48,
+    0.60,
+    0.75,
+    0.81,
+    0.77,
+    0.73,
+    0.69,
+    0.65,
+    0.61,
+    0.57,
+    0.53,
+    0.49,
+    0.45,
+    0.41,
+    0.37,
+    0.33,
+    0.29,
+    0.25,
+    0.21,
+    0.36,
+    0.51,
+    0.66,
+    0.81,
+    0.77,
+    0.73,
+    0.85,
+    0.81,
+]
 
 # Define the conversion factor
 in_to_pa = 248.84
 
 # Convert to Pascals
-data["setpoint"]  = [value * in_to_pa for value in values_in_inches]
-data["number_of_requests"] = [0, 1, 2, 3, 4, 6, 3, 0, 0, 0, 1, 1, 1, 1, 2, 2, 2, 2, 1, 0, 1, 2, 6, 6, 5, 5, 2, 2, 4, 2]
+data["setpoint"] = [value * in_to_pa for value in values_in_inches]
+data["number_of_requests"] = [
+    0,
+    1,
+    2,
+    3,
+    4,
+    6,
+    3,
+    0,
+    0,
+    0,
+    1,
+    1,
+    1,
+    1,
+    2,
+    2,
+    2,
+    2,
+    1,
+    0,
+    1,
+    2,
+    6,
+    6,
+    5,
+    5,
+    2,
+    2,
+    4,
+    2,
+]
 data["flag_device"] = [1 for _ in range(30)]
-# fmt: on
 
 
 class TestTrimRespond(unittest.TestCase):
@@ -352,29 +413,83 @@ class TestTrimRespond(unittest.TestCase):
     def test_TR_winter_day_reset_singlesample(self):
         """Test the winter day reset dataset from Modelica"""
 
-        start_date = "2023-12-21 12:00:00"
         modelica_winter = pd.DataFrame(
             columns=["setpoint", "number_of_requests", "flag_device"],
-            index=pd.date_range(start=start_date, periods=1441, freq="min"),
+            index=pd.date_range(start="2023-12-21 12:00:00", periods=1441, freq="min"),
         )
-        # fmt: off
+
         modelica_winter["setpoint"] = (
             [120] * 294
             + [
-                108, 96, 84, 72, 60, 48, 51, 71, 91, 111, 131, 151, 171, 191, 211, 231,
-                251, 271, 291, 311,  314, 317, 305, 293, 281, 269, 257, 245, 233, 221,
[…]
