### 2025-06-04 a2a9e55
Added a docstring
