+++ constrain/library/__init__.py
@@ -1,6 +1,10 @@
 from .AppendixGHVACSystemFanOperation import *
 from .automatic_oa_damper_controls import *
 from .automatic_shutdown import *
+from .chilled_water_plant_sizing_chiller_max_loading import *
+from .chilled_water_plant_sizing_chiller_short_cycling import *
+from .chilled_water_plant_sizing_multiple_chillers import *
+from .chilled_water_plant_sizing_whole_plant import *
 from .demand_control_vent import *
 from .ExteriorLightingControlDaylightOff import *
 from .ExteriorLightingControlOccupancySensingReduction import *
@@ -116,4 +120,8 @@ __all__ = [
     "G36ReheatTerminalBoxHeatingAirflowSetpoint",
     "G36ReheatTerminalBoxDeadbandAirflowSetpoint",
     "G36TerminalBoxCoolingMinimumAirflow",
+    "ChilledWaterPlantSizingWholePlant",
+    "ChilledWaterPlantSizingMultipleChillers",
+    "ChilledWaterPlantSizingChillerMaxLoading",
+    "ChilledWaterPlantSizingChillerShortCycling",
 ]

[…]
