+++ constrain/library/chilled_water_plant_sizing_whole_plant.py
@@ -66,11 +66,11 @@ from constrain.checklib import RuleCheckBase
 from sklearn.linear_model import LinearRegression
 from sklearn.metrics import r2_score
 from sklearn.preprocessing import PolynomialFeatures
+import numpy as np
 
 
-Cp = 4.184  # kJ/(kg*K), specific heat
+Cp = 4184  # J/(kg*K)
 UPPER_TEMP_AIR_OUTDOOR_LIMIT = 48.89  # deg C ~= 120 deg F
-ACCEPTABLE_CHILLER_SIZING_RATIO = 0.95
 
 
 class ChilledWaterPlantSizingWholePlant(RuleCheckBase):
@@ -81,10 +81,24 @@ class ChilledWaterPlantSizingWholePlant(RuleCheckBase):
         "load_plant_water_chilled",
         "temperature_air_outdoor",
         "capacity_nominal_plant_water_chilled",
+        "ratio_sizing_plant_water_chilled_acceptable",
     ]
 
-    def verify(self):
+    def calc_temp_from_load(self, load, model):
+        a = model.coef_[2]
+        b = model.coef_[1]
+        c = model.intercept_ - load
 
+        discriminant = b**2 - 4 * a * c
+        if discriminant < 0:
+            raise ValueError("No real solution exists for the given y.")
+
+        x1 = (-b + np.sqrt(discriminant)) / (2 * a)
+        x2 = (-b - np.sqrt(discriminant)) / (2 * a)
+
+        return max(x1, x2)
+
+    def verify(self):
         # Calculate `load_plant_water_chilled` if any value is None
         if self.df["load_plant_water_chilled"].isnull().any():
             self.df["load_plant_water_chilled"] = (
@@ -96,83 +110,36 @@ class ChilledWaterPlantSizingWholePlant(RuleCheckBase):
                 )
             )
 
-        # Perform linear, quadratic, cubic  regressions of `load_plant_water_chilled` = f(`temperature_air_outdoor`)
-        # 1. linear
-        linear_model = LinearRegression().fit(
-            self.df[["temperature_air_outdoor"]], self.df[["load_plant_water_chilled"]]
+        # Perform regressions of `load_plant_water_chilled` = f(`temperature_air_outdoor`)
+        feat = PolynomialFeatures(degree=2)
+        X = feat.fit_transform(self.df[["temperature_air_outdoor"]].values)
+        model = LinearRegression().fit(X, self.df["load_plant_water_chilled"])
+        y = model.predict(X)
+        coefficient_of_determination = r2_score(
+            self.df[["load_plant_water_chilled"]], y
         )
-
-        # 1.1 Make a prediction
-        y_pred_linear = linear_model.predict(self.df[["temperature_air_outdoor"]])
-
-        # 1.2 Calculate R^2 value
-        r_squared_linear = r2_score(
-            self.df[["load_plant_water_chilled"]], y_pred_linear
-        )
-
-        # 2. quadratic
-        # 2.1 Transform features to quadratic (second-degree)
-        quadratic_feature = PolynomialFeatures(degree=2)
-        x_quadratic = quadratic_feature.fit_transform(
-            self.df[["temperature_air_outdoor"]]
-        )
-
-        # 2.2 train the model
-        quadratic_model = LinearRegression().fit(
-            x_quadratic, self.df[["load_plant_water_chilled"]]
-        )
-
-        # 2.3 Make a prediction
-        y_quadratic_pred = quadratic_model.predict(x_quadratic)
-
-        # 2.4 Calculate R^2 value
-        r_squared_quadratic = r2_score(
-            self.df[["load_plant_water_chilled"]], y_quadratic_pred
-        )
-
-        # 3. cubic
[…]
