### 2025-06-19 57dbf71
Filter null load points.
