### 2025-06-20 ea391e7
Revise transition of operation and base verification on rolling hour.
