+++ tests/test_chilled_water_plant_sizing_chiller_short_cycling.py
@@ -9,7 +9,7 @@ from library import *
 
 
 class TestChilledWaterPlantSizingChillerShortCycling(unittest.TestCase):
-    def test_chilled_water_plant_sizing_chiller_short_cycling_power_input_pass(
+    def test_chilled_water_plant_sizing_chiller_short_cycling_pass(
         self,
     ):
         points = [
@@ -17,47 +17,41 @@ class TestChilledWaterPlantSizingChillerShortCycling(unittest.TestCase):
             "cycles_number_maximum",
         ]
         timestamp = [
-            datetime(2023, 11, 1, 8, 0),
-            datetime(2023, 11, 1, 8, 15),
-            datetime(2023, 11, 1, 8, 30),
-            datetime(2023, 11, 1, 8, 45),
             datetime(2023, 11, 1, 9, 0),
-            datetime(2023, 11, 1, 9, 15),
-            datetime(2023, 11, 1, 9, 30),
+            datetime(2023, 11, 1, 9, 35),
             datetime(2023, 11, 1, 9, 45),
             datetime(2023, 11, 1, 10, 0),
             datetime(2023, 11, 1, 10, 15),
-            datetime(2023, 11, 1, 10, 30),
-            datetime(2023, 11, 1, 10, 45),
-            datetime(2023, 11, 1, 11, 00),
+            datetime(2023, 11, 1, 11, 15),
+            datetime(2023, 11, 1, 11, 16),
+            datetime(2023, 11, 1, 11, 17),
+            datetime(2023, 11, 1, 11, 20),
+            datetime(2023, 11, 1, 11, 25),
+            datetime(2023, 11, 1, 11, 59),
         ]
         data = [
-            [0, 2],
-            [0, 2],
-            [0, 2],
-            [0, 2],
-            [0, 2],
-            [0, 2],
-            [0, 2],
-            [5.0, 2],
-            [5.0, 2],
-            [4.8, 2],
-            [4.7, 2],
-            [4.5, 2],
-            [5.0, 2],
+            [0, 3],
+            [1, 3],
+            [1, 3],
+            [0, 3],
+            [1, 3],
+            [0, 3],
+            [0, 3],
+            [1, 3],
+            [1, 3],
+            [1, 3],
+            [0, 3],
         ]
         df = pd.DataFrame(data, columns=points, index=timestamp)
-        expected_results = pd.Series([True, True, True, True])
         verification_obj = run_test_verification_with_data(
             "ChilledWaterPlantSizingChillerShortCycling", df
         )
 
         results = pd.Series(list(verification_obj.result))
         binaryflag = verification_obj.check_bool()
-        self.assertTrue(results.equals(expected_results))
         self.assertTrue(binaryflag)
 
-    def test_chilled_water_plant_sizing_chiller_short_cycling_all_hours_less_than_max_cycling_pass(
+    def test_chilled_water_plant_sizing_chiller_short_cycling_fail(
         self,
     ):
         points = [
@@ -65,98 +59,42 @@ class TestChilledWaterPlantSizingChillerShortCycling(unittest.TestCase):
             "cycles_number_maximum",
         ]
         timestamp = [
-            datetime(2023, 11, 1, 8, 0),
-            datetime(2023, 11, 1, 8, 15),
-            datetime(2023, 11, 1, 8, 30),
-            datetime(2023, 11, 1, 8, 45),
             datetime(2023, 11, 1, 9, 0),
-            datetime(2023, 11, 1, 9, 15),
-            datetime(2023, 11, 1, 9, 30),
+            datetime(2023, 11, 1, 9, 35),
             datetime(2023, 11, 1, 9, 45),
             datetime(2023, 11, 1, 10, 0),
             datetime(2023, 11, 1, 10, 15),
-            datetime(2023, 11, 1, 10, 30),
[…]
