### 2025-06-25 0b37e31
Revise verifications, clean up, update unit tests.
