+++ constrain/library/chilled_water_plant_sizing_whole_plant.py
@@ -24,30 +24,17 @@ We aim to verify that the chilled water plant is sized appropriately for the des
 
 ### Verification Algorithm Pseudo Code
 
-The algorithm performs regression analysis to predict peak cooling load based on outdoor air temperature:
-
-1. Calculate chilled water plant load if not provided directly
-2. Perform linear, quadratic, and cubic regressions of plant load vs. outdoor air temperature
-3. Select the regression model with the best fit (highest R² value)
-4. Use the selected model to predict the peak load at the maximum outdoor air temperature
-5. Compare the predicted peak load with the nominal plant capacity
-6. Verification passes if:
-   - The maximum outdoor air temperature doesn't exceed the upper limit (120°F/48.89°C)
-   - The predicted peak load is less than or equal to 95% of the nominal plant capacity
+```
+if temperature_drybulb_day_design_cooling is in dataset
+    if peak_load >= nominal_chilled_water_plant_capacity / oversizing_factor
+        return True
+    else
+        return False
+else
+    return "Untested"
+```
 
 ### Data requirements
-
-- temperature_water_supply: Chilled water supply temperature
-  - Data Value Unit: °C
-  - Data Point Affiliation: Chilled water plant
-
-- temperature_water_return: Chilled water return temperature
-  - Data Value Unit: °C
-  - Data Point Affiliation: Chilled water plant
-
-- flow_mass_water: Chilled water mass flow rate
-  - Data Value Unit: kg/s
-  - Data Point Affiliation: Chilled water plant
 
 - load_plant_water_chilled: Chilled water plant cooling load
   - Data Value Unit: kW
@@ -60,94 +47,53 @@ The algorithm performs regression analysis to predict peak cooling load based on
 - capacity_nominal_plant_water_chilled: Nominal capacity of the chilled water plant
   - Data Value Unit: kW
   - Data Point Affiliation: Chilled water plant
+
+- temperature_drybulb_day_design_cooling: Cooling design day drybulb temperature
+  - Data Value Unit: °C
+  - Data Point Affiliation: Design parameter
+
+- factor_oversizing: Oversizing factor
+  - Data Value Unit: fraction
+  - Data Point Affiliation: Design parameter
 """
 
 from constrain.checklib import RuleCheckBase
-from sklearn.linear_model import LinearRegression
-from sklearn.metrics import r2_score
-from sklearn.preprocessing import PolynomialFeatures
-import numpy as np
-
 
 Cp = 4184  # J/(kg*K)
-UPPER_TEMP_AIR_OUTDOOR_LIMIT = 48.89  # deg C ~= 120 deg F
 
 
 class ChilledWaterPlantSizingWholePlant(RuleCheckBase):
     points = [
-        "temperature_water_supply",
-        "temperature_water_return",
-        "flow_mass_water",
         "load_plant_water_chilled",
         "temperature_air_outdoor",
         "capacity_nominal_plant_water_chilled",
-        "ratio_sizing_plant_water_chilled_acceptable",
-        "determination_of_coefficient_acceptable",
+        "temperature_drybulb_day_design_cooling",
+        "factor_oversizing",
     ]
 
-    def calc_temp_from_load(self, load, model):
-        a = model.coef_[2]
-        b = model.coef_[1]
-        c = model.intercept_ - load
-
-        discriminant = b**2 - 4 * a * c
-        if discriminant < 0:
-            raise ValueError("No real solution exists for the given y.")
-
-        x1 = (-b + np.sqrt(discriminant)) / (2 * a)
-        x2 = (-b - np.sqrt(discriminant)) / (2 * a)
-
-        return max(x1, x2)
-
     def verify(self):
[…]
