+++ .github/workflows/unit_tests.yml
@@ -37,6 +37,6 @@ jobs:
           path: ./.venv
           key: venv-${{ hashFiles('poetry.lock') }}
       - name: Install the project dependencies
-        run: poetry install -- with dev
+        run: poetry install --with dev
       - name: Run the tests
         run: poetry run pytest -v
