### 2025-07-09 af6f129
Fixed brick class name in doc
