### 2025-07-24 d154355
Modify create/edit state dialog to be scrollable
