+++ constrain/checklib.py
@@ -1,4 +1,6 @@
 """
+checklib.py
+====================================
 This file containing the high level interface for implementing verification item classes in library.py
 """
 
