+++ docs/source/Quickstart Guide.rst
@@ -10,6 +10,12 @@ Quickstart Guide
 
 This guide provides a quick overview of how to get started with **ConStrain**. Verifications of building system control related timeseries using **ConStrain** can either be done by using **ConStrain** as a Python library or by using a **ConStrain** workflow.
 
+What? Who? How?
+------------------
+**ConStrain** is a data-driven knowledge-integrated framework that automatically verifies that controls function as intended. As such, **ConStrain** might be of interest to a wide audience: from commissioning agents, to building operators, building energy modeling software users and developers and more. 
+
+The first step towards using the tool is to gather timeseries data that can be used to verify a specific sequence of operation or the correct behavior of an equipment. **ConStrain**'s library_ provide a list of the different verifications that can be performed, not that they can it can be expanded see documentation section on that topic. For each, it provides some general description and references, and also provides a list of the data points needed to carry out verifications. For example, the ASHRAE Guideline 36 verification named `G36ReheatTerminalBoxDeadbandAirflowSetpoint` aims to very that terminal boxes with reheat when the zone state is deadband following the Guideline 36 recommendations as described in Section 5.6.5.2 in ASHRAE Guideline 36 2021. For this verification to be carried out, a user would have to gather timeseries (or static parameter values when applicable) for the following datapoints: system operation mode, zone state, minimum airflow setpoint during occupied mode, airflow setpoint, heating coil command, discharge air temperature, minimum discharge air temperature setpoint. Note that the timeseries can come from any sources: from a building management system or from simulation. When completed, one is ready to use ConStrain. Note that if the data needs to be pre-processed, **ConStrain**'s pre-processing API_ could help.
+
 Installing **ConStrain**
 -------------------------
 **ConStrain** can be installed from PyPI by running the following command.
@@ -24,8 +30,8 @@ Running Verifications using **ConStrain**
 **ConStrain** relies on verification case files to perform verifications. These are JSON files that contain all the necessary information about data points, reference to verification logics, and simulation parameters when applicable. A verification case file can be built as follows:
 
   1. Identify if the desired verification is already part of the default library. **ConStrain**'s default verifications are documented :doc:`here <../Verifications>`. If not, consider expanding the default verifications, help is provided :doc:`here <../Expand Exisiting Verifications>`.
-  2. Create a JSON file that contains all the information needed by ConStrain to run the verification as detailed below or as defined on this schema (TBD).
-  3. Run the verification either using **ConStrain** as a :ref:`library <library>` or using a **ConStrain** :ref:`workflow <workflow>`.
+  2. Create a JSON file that contains all the information needed by ConStrain to run the verification as detailed below or as defined in this schema_.
+  3. Run the verification either using **ConStrain** as a :ref:`library <lib>` or using a **ConStrain** :ref:`workflow <wf>`.
 
 .. sourcecode:: JSON
 
@@ -74,7 +80,7 @@ Running Verifications using **ConStrain**
     }
   }
 
-.. _library:
+.. _lib:
 
 Using **ConStrain** as a Python Library
 ----------------------------------------
@@ -138,191 +144,82 @@ Finally, we can create summary report. A summary report for all verification wil
 
     reporting.report_multiple_cases()
 
-.. _workflow:
+.. _wf:
 
-Using **ConStrain**'s' Workflows
+Using **ConStrain**'s Workflows
 ----------------------------------
 
[…]

[…]
