### 2025-07-25 c742dae
Add images to README
