### 2025-07-25 e82fb1b
Black formatting
