### 2025-07-28 69dd4b6
change schema term names
