### 2025-07-28 2a2da22
update paper contents based on reviewer comments
