+++ docs/paper.bib
@@ -54,3 +54,45 @@
     author={{U.S. Department of Energy}},
     howpublished={\url{https://www.energy.gov/eere/articles/decarbonizing-us-economy-2050}},
     note={Accessed:2024-07-30}}
+
+@misc{skyspark,
+  author = {SkyFoundry},
+  title = {SkySpark Analytics},
+  year = {2023},
+  note = {\url{https://skyfoundry.com/skyspark}},
+}
+
+@misc{coppertree,
+  author = {CopperTree Analytics},
+  title = {CopperTree Analytics Platform},
+  year = {2023},
+  note = {\url{https://www.coppertreeanalytics.com}},
+}
+
+@article{fdd_review,
+  author = {Katipamula, Srinivas and Brambley, Michael R.},
+  title = {Review of Fault Detection and Diagnostic Methods for Building Systems},
+  journal = {HVAC&R Research},
+  volume = {11},
+  number = {2},
+  pages = {169--187},
+  year = {2005},
+  doi = {10.1080/10789669.2005.10391123}
+}
+
+@inproceedings{brick,
+  author = {Balaji, Bharathan and Bhattacharya, Arka and Fierro, Gabe and Gao, Jie and Gluck, Josh and Hong, Deokwoo and Johansen, Aslak and Koh, Jihyun and Ploennigs, Joern and Agarwal, Yuvraj and Whitehouse, Kamin and Taneja, Jayant},
+  title = {Brick: Metadata Schema for Portable Smart Building Applications},
+  booktitle = {Proceedings of the 3rd ACM International Conference on Systems for Energy-Efficient Built Environments},
+  year = {2016},
+  pages = {41--50},
+  publisher = {ACM},
+  doi = {10.1145/2993422.2993577}
+}
+
+@manual{energyplus,
+  title        = {EnergyPlus Energy Simulation Software},
+  author       = {{U.S. Department of Energy}},
+  organization = {U.S. Department of Energy},
+  url          = {https://energyplus.net/}
+}
