### 2025-07-28 c804422
add back authors
