### 2025-07-29 4e3be02
Update paper.md
Generalized language by removing company-specific tool names and refining wording to focus on industry practices and performance-based control verification
