### 2025-07-29 44e3d00
Add reference to the new Brick How-to guide.
