+++ tests/api/test_verification_case.py
@@ -1,10 +1,10 @@
 import json
 import os, sys, unittest, copy
 
+# Add the project root to the path so we can import constrain
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
 
-sys.path.append("./constrain")
-
-from api import VerificationCase
+from constrain.api import VerificationCase
 
 
 class TestVerificationCase(unittest.TestCase):
