+++ tests/api/test_brick_compliance.py
@@ -1,9 +1,11 @@
 import sys
+import os
 import unittest
 
-sys.path.append("./constrain")
-from api import BrickCompliance
+# Add the project root to the path so we can import constrain
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
 
+from constrain.api import BrickCompliance
 
 class TestBrickCompliance(unittest.TestCase):
     def test_constructor(self):
