### 2025-07-29 7e9b7aa
Black formatting.
