+++ tests/api/test_brick_compliance.py
@@ -3,9 +3,10 @@ import os
 import unittest
 
 # Add the project root to the path so we can import constrain
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from constrain.api import BrickCompliance
+
 
 class TestBrickCompliance(unittest.TestCase):
     def test_constructor(self):
