+++ tests/api/test_verification_library.py
@@ -1,7 +1,7 @@
 import unittest, sys, json, os
 
 # Add the project root to the path so we can import constrain
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from constrain.api import VerificationLibrary
 
