+++ tests/api/test_flexible_calling.py
@@ -3,7 +3,7 @@ import platform
 from unittest.mock import patch
 
 # Add the project root to the path so we can import constrain
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from constrain.api import Workflow
 
