+++ tests/api/test_workflow.py
@@ -4,7 +4,7 @@ import sys
 import os
 
 # Add the project root to the path so we can import constrain
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from constrain.api import Workflow
 
