+++ tests/api/test_verification.py
@@ -3,7 +3,7 @@ import shutil
 import unittest, sys, os, pathlib
 
 # Add the project root to the path so we can import constrain
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', '..'))
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "..", ".."))
 
 from constrain.api import VerificationCase
 from constrain.api import Verification
