### 2025-07-29 4392073
change global variable names
