+++ tests/api/test_reporting.py
@@ -5,17 +5,17 @@ import unittest
 sys.path.append("./constrain")
 from api import Reporting
 
-verification_json = "./tests/api/data/verification_output/*_md.json"
-result_md_name = "testing.md"
-result_md_path = "./tests/api/data/verification_output/testing.md"
-report_format = "markdown"
+VERIFICATION_JSON = "./tests/api/data/verification_output/*_md.json"
+RESULT_MD_NAME = "testing.md"
+RESULT_MD_PATH = "./tests/api/data/verification_output/testing.md"
+REPORT_FORMAT = "markdown"
 
 
 class TestReporting(unittest.TestCase):
     def test_constructor_wrong_arg_type(self):
         # wrong `verification_json` type
         with self.assertLogs() as logobs:
-            Reporting({verification_json}, result_md_name, report_format)
+            Reporting({VERIFICATION_JSON}, RESULT_MD_NAME, REPORT_FORMAT)
             self.assertEqual(
                 "ERROR:root:The type of the `verification_json` arg needs to be a str. It cannot be <class 'set'>.",
                 logobs.output[0],
@@ -23,7 +23,7 @@ class TestReporting(unittest.TestCase):
 
         # wrong `result_md_path` type
         with self.assertLogs() as logobs:
-            Reporting(verification_json, [result_md_name], report_format)
+            Reporting(VERIFICATION_JSON, [RESULT_MD_NAME], REPORT_FORMAT)
             self.assertEqual(
                 "ERROR:root:The type of the `result_md_name` arg needs to be a str. It cannot be <class 'list'>.",
                 logobs.output[0],
@@ -31,27 +31,27 @@ class TestReporting(unittest.TestCase):
 
         # wrong `report_format` type
         with self.assertLogs() as logobs:
-            Reporting(verification_json, result_md_name, [report_format])
+            Reporting(VERIFICATION_JSON, RESULT_MD_NAME, [REPORT_FORMAT])
             self.assertEqual(
                 "ERROR:root:The type of the `report_format` arg needs to be a str. It cannot be <class 'list'>.",
                 logobs.output[0],
             )
 
     def test_report_multiple_cases(self):
-        reporting_obj = Reporting(verification_json, result_md_name, report_format)
+        reporting_obj = Reporting(VERIFICATION_JSON, RESULT_MD_NAME, REPORT_FORMAT)
 
         # report only selective verification results
         reporting_obj.report_multiple_cases(item_names=["AutomaticOADamperControl"])
-        self.assertTrue(os.path.isfile(result_md_path))
-        os.remove(result_md_path)
+        self.assertTrue(os.path.isfile(RESULT_MD_PATH))
+        os.remove(RESULT_MD_PATH)
 
         # report all the verification results
         reporting_obj.report_multiple_cases(item_names=[])
-        self.assertTrue(os.path.isfile(result_md_path))
-        os.remove(result_md_path)
+        self.assertTrue(os.path.isfile(RESULT_MD_PATH))
+        os.remove(RESULT_MD_PATH)
 
     def test_report_multiple_cases_wrong_arg_type(self):
-        reporting_obj = Reporting(verification_json, result_md_name, report_format)
+        reporting_obj = Reporting(VERIFICATION_JSON, RESULT_MD_NAME, REPORT_FORMAT)
 
         # wrong `item_names` type
         with self.assertLogs() as logobs:
@@ -64,7 +64,7 @@ class TestReporting(unittest.TestCase):
             )
 
     def test_report_multiple_cases_wrong_verification_name(self):
-        reporting_obj = Reporting(verification_json, result_md_name, report_format)
+        reporting_obj = Reporting(VERIFICATION_JSON, RESULT_MD_NAME, REPORT_FORMAT)
 
         # wrong verification item name
         with self.assertLogs() as logobs:
