### 2025-07-29 14f0470
remove print for the rest, update a couple of logging type
