+++ constrain/library/G36ReheatTerminalBoxHeatingCoilTracking.py
@@ -66,6 +66,7 @@ else:
   - Data Point Affiliation: Terminal box control
   
 """
+import logging
 
 import pandas as pd
 from constrain.checklib import RuleCheckBase
@@ -134,7 +135,7 @@ class G36ReheatTerminalBoxHeatingCoilTracking(RuleCheckBase):
                     else:
                         result_flag = False
                 else:
-                    print("invalid error time")
+                    logging.error("invalid error time")
                     return False
 
             self.result.loc[cur_time] = result_flag
