+++ constrain/library/G36TerminalBoxVAVDamperTracking.py
@@ -60,6 +60,7 @@ else:
   - Data Point Affiliation: Terminal box control
 
 """
+import logging
 
 import pandas as pd
 from constrain.checklib import RuleCheckBase
@@ -122,7 +123,7 @@ class G36TerminalBoxVAVDamperTracking(RuleCheckBase):
                 else:
                     result_flag = False
             else:
-                print("invalid error time")
+                logging.error("invalid error time")
                 return False
 
             self.result.loc[cur_time] = result_flag
