+++ constrain/library/G36CoolingOnlyTerminalBoxCoolingAirflowSetpoint.py
@@ -67,6 +67,7 @@ end
   - Data Point Affiliation: Zone airflow control
 
 """
+import logging
 
 from constrain.checklib import RuleCheckBase
 
@@ -94,7 +95,7 @@ class G36CoolingOnlyTerminalBoxCoolingAirflowSetpoint(RuleCheckBase):
                 cooling_maximum = 0
                 cooling_minimum = 0
             case _:
-                print("invalid operation mode value")
+                logging.info("invalid operation mode value")
                 return "Untested"
 
         if cooling_minimum <= v_sp <= cooling_maximum:

[…]
