+++ constrain/library/heat_rejection_fan_var_flow_control.py
@@ -73,6 +73,7 @@ else:
   - Data Point Affiliation: Equipment specifications
 
 """
+
 import logging
 from typing import Dict
 
