+++ constrain/library/guest_room_control_vent.py
@@ -73,6 +73,7 @@ for each day:
   - Data Point Affiliation: Zone ventilation
 
 """
+
 import logging
 
 import pandas as pd
