+++ constrain/library/G36CoolingOnlyTerminalBoxDeadbandAirflowSetpoint.py
@@ -58,6 +58,7 @@ end
   - Data Point Affiliation: Zone airflow control
 
 """
+
 import logging
 
 from constrain.checklib import RuleCheckBase
