+++ constrain/library/G36TerminalBoxCoolingMinimumAirflow.py
@@ -71,6 +71,7 @@ else:
   - Data Point Affiliation: Zone monitoring
   
 """
+
 import logging
 
 from constrain.checklib import RuleCheckBase
