### 2025-07-29 eee1128
test assert refactor test
