+++ tests/api/test_verification_case.py
@@ -236,13 +236,12 @@ class TestVerificationCase(unittest.TestCase):
             suite_keys = list(vc.case_suite.keys())
             partial_good_keys = suite_keys[:2] + ["wrong-key-1", 2]
             vc.save_case_suite_to_json(saving_file_path, case_ids=partial_good_keys)
-            self.assertEqual(
-                "WARNING:root:case_id wrong-key-1 is not in self.case_suite!",
-                logobs.output[0],
+            self.assertTrue(
+                "WARNING:root:case_id wrong-key-1 is not in self.case_suite!"
+                in logobs.output
             )
-            self.assertEqual(
-                "WARNING:root:case_id 2 is not in self.case_suite!",
-                logobs.output[1],
+            self.assertTrue(
+                "WARNING:root:case_id 2 is not in self.case_suite!" in logobs.output
             )
 
         with open(saving_file_path, "r") as f:
