+++ tests/api/test_brick_compliance.py
@@ -16,9 +16,9 @@ class TestBrickCompliance(unittest.TestCase):
                 brick_schema_path=["./resources/brick/Brick.ttl"],
                 brick_instance_path="./resources/brick/brick_instance.ttl",
             )
-            self.assertEqual(
-                "ERROR:root:The `brick_schema_path` argument type must be str, but <class 'list'> type is provided.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The `brick_schema_path` argument type must be str, but <class 'list'> type is provided."
+                in logobs.output
             )
 
         # test `brick_instance_path` type
@@ -27,9 +27,9 @@ class TestBrickCompliance(unittest.TestCase):
                 brick_schema_path="./resources/brick/Brick.ttl",
                 brick_instance_path=["./resources/brick/brick_instance.ttl"],
             )
-            self.assertEqual(
-                "ERROR:root:The `brick_instance_path` argument type must be str, but <class 'list'> type is provided.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The `brick_instance_path` argument type must be str, but <class 'list'> type is provided."
+                in logobs.output
             )
 
         # test `datapoint_name_conversion_path` type
@@ -41,9 +41,9 @@ class TestBrickCompliance(unittest.TestCase):
                     "./resources/brick/query_statement.yml"
                 ],
             )
-            self.assertEqual(
-                "ERROR:root:The `datapoint_name_conversion_path` argument type must be str, but <class 'list'> type is provided.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The `datapoint_name_conversion_path` argument type must be str, but <class 'list'> type is provided."
+                in logobs.output
             )
 
         # test `datapoint_name_conversion_path` type
@@ -55,9 +55,9 @@ class TestBrickCompliance(unittest.TestCase):
                     "./resources/brick/verification_datapoint_info.yml"
                 ],
             )
-            self.assertEqual(
-                "ERROR:root:The `datapoint_name_conversion_path` argument type must be str, but <class 'list'> type is provided.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The `datapoint_name_conversion_path` argument type must be str, but <class 'list'> type is provided."
+                in logobs.output
             )
 
         # test `perform_reasoning` type
@@ -67,9 +67,9 @@ class TestBrickCompliance(unittest.TestCase):
                 brick_instance_path="./resources/brick/brick_instance.ttl",
                 perform_reasoning="False",
             )
-            self.assertEqual(
-                "ERROR:root:The `perform_reasoning` argument type must be bool, but <class 'str'> type is provided.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The `perform_reasoning` argument type must be bool, but <class 'str'> type is provided."
+                in logobs.output
             )
 
         # test if `brick_schema_path` exists
@@ -78,9 +78,9 @@ class TestBrickCompliance(unittest.TestCase):
                 brick_schema_path="./wrong_path.ttl",
                 brick_instance_path="./resources/brick/brick_instance.ttl",
             )
-            self.assertEqual(
-                "ERROR:root:The file doesn't exist in the provided directory. Please make sure to provide a correct `brick_schema_path`.",
-                logobs.output[0],
+            self.assertTrue(
+                "ERROR:root:The file doesn't exist in the provided directory. Please make sure to provide a correct `brick_schema_path`."
+                in logobs.output
             )
 
         # test if `brick_instance_path` exists
@@ -89,9 +89,9 @@ class TestBrickCompliance(unittest.TestCase):
                 brick_schema_path="./resources/brick/Brick.ttl",
                 brick_instance_path="./wrong.ttl",
[…]
