### 2025-07-29 8095944
mod lib base class property
