+++ constrain/checklib.py
@@ -22,11 +22,12 @@ class CheckLibBase(ABC):
     """Abstract class defining interfaces for item-specific verification classes"""
 
     points = None
-    result = pd.DataFrame()
 
     def __init__(
         self, df: pd.DataFrame, params=None, results_folder=None, tolerances=None
     ):
+        self.result = pd.DataFrame()
+        self.points_list = self.points
         full_df = df.copy(deep=True)
         if params is not None:
             for k, v in params.items():
@@ -47,9 +48,6 @@ class CheckLibBase(ABC):
         self.result.name = ""
         self.df["Verification Result"] = self.result
 
-    @property
-    def points_list(self) -> List[str]:
-        return self.points
 
     @abstractmethod
     def check_bool(self) -> bool:
