+++ docs/paper.bib
@@ -96,3 +96,24 @@
   organization = {U.S. Department of Energy},
   url          = {https://energyplus.net/}
 }
+
+@article{spawn,
+  title={Spawn: coupling Modelica Buildings Library and EnergyPlus to enable new energy system and control applications},
+  author={Wetter, Michael and Benne, Kyle and Tummescheit, Hubertus and Winther, Christian},
+  journal={Journal of Building Performance Simulation},
+  volume={17},
+  number={2},
+  pages   = {274-292},
+  year    = {2024},
+  publisher = {Taylor \& Francis},
+  doi     = {10.1080/19401493.2023.2266414},
+  url     = {https://www.tandfonline.com/doi/full/10.1080/19401493.2023.2266414}
+}
+
+@misc{modelica,
+  author       = {Modelica Association},
+  title        = {Modelica – A Unified Object-Oriented Language for Physical Systems Modeling},
+  year         = {2023},
+  note         = {Modelica Specification Version 3.6},
+  url          = {https://specification.modelica.org/maint/3.6/MLS.pdf}
+}
