+++ docs/paper.md
@@ -49,14 +49,14 @@ Advances in building control have shown significant potential for improving buil
 
 However, one of the challenges to realizing those savings is the correct implementation of such advanced control strategies and regularly verifying their actual operational performance. A field study found that only 50% of systems observed have their control system correctly configured to meet the energy codes requirement [@impl_ctrl], and control-related compliance verification is typically not included in the commissioning scope.
 
-Current control verification is often conducted manually, which is time-consuming, ad-hoc, incomplete, and error-prone. To address this, `ConStrain` focuses on formalizing and automating verification of HVAC controls by analyzing sensor and actuator data streams from building control systems. `ConStrain` is an open-source library and application programming interface (API) for analyzing BAS data streams for adherence to an operational specification, which can correspond to building energy code (e.g., ASHRAE 90.1) or to a high-performance building control guideline (e.g., ASHRAE Guideline 36 [@g36]). `ConStrain` is also incorporating semantic modeling [@brick] capabilities to enable automated configuration and deployment of verification.
+Current control verification is often conducted manually, which is time-consuming, ad-hoc, incomplete, and error-prone. To address this, `ConStrain` focuses on formalizing and automating verification of HVAC controls by analyzing sensor and actuator data streams from building control systems. `ConStrain` is an executable control verification knowledge base and application programming interface (API) for analyzing BAS data streams for adherence to an operational specification, which can correspond to building energy code (e.g., ASHRAE 90.1) or to a high-performance building control guideline (e.g., ASHRAE Guideline 36 [@g36]). `ConStrain` is also incorporating semantic modeling [@brick] capabilities to enable automated configuration and deployment of verification.
 
-`ConStrain` can be used as a standalone tool and can also be integrated into established workflows of third-party tools and practices. For instance, `ConStrain` has been successfully integrated as part of the continuous integration software development process of whole-building energy simulation-based software tool (e.g., Washington State's Total System Performance Ratio Analysis Tool [@tspr]) to make sure that software code contributions as well as simulation software updates do not have unexpected impacts on the simulated performance of building system controls. Moreover, a set of `OpenStudio` [@os] measures [@osm] have also been developed to enable building energy modelers using `OpenStudio` to have access to perform verification on their models with minimal configurations required.
+`ConStrain` can be used as a standalone tool and can also be integrated into established workflows of third-party tools. For instance, `ConStrain` has been successfully integrated as part of the continuous integration software development process of whole-building energy simulation-based software tool (e.g., Washington State's Total System Performance Ratio Analysis Tool [@tspr]) to make sure that software code contributions as well as simulation software updates do not have unexpected impacts on the simulated performance of building system controls. Moreover, a set of `OpenStudio` [@os] measures [@osm] have also been developed to enable building energy modelers using `OpenStudio` to perform verification on their models with minimal configurations required.
 
 
 ### Comparison with existing tools and industry practices
 
-In current industry practices, HVAC control verification is often conducted manually by commissioning agents or facilities teams, or through proprietary trend data analytics solutions integrated into Building Automation Systems (BAS). These tools, while valuable for fault detection and system monitoring, are generally vendor-specific, offer limited transparency, and are not purpose-built to ensure that control strategies intended to deliver energy savings, such as those specified in energy codes or design standards, are functioning as intended in actual operation. 
[…]
