+++ docs/paper.bib
@@ -17,9 +17,10 @@
     }
 
 @misc{tspr,
-    title={TSPR Washington State Analysis Tool},
+    title={TSPR Washington State Analysis Tool 2024.1.0},
     author={{Pacific Northwest National Laboratory}},
     howpublished={\url{https://energycode.pnl.gov/HVACSystemPerformance/}},
+    year={2024},
     note={Accessed: 2024-04-044}
 }
 
@@ -36,9 +37,10 @@
 }
 
 @misc{os,
-    title={OpenStudio},
+    title={OpenStudio 3.8},
     author={{National Renewable Energy Laboratory, Argonne National Laboratory, Lawrence Berkeley National Laboratory, Oak Ridge National Laboratory, Pacific Northwest National Laboratory}},
     howpublished={\url{https://openstudio.net/}},
+    year={2024},
     note={Accessed: 2024-04-04}
 }
 
@@ -46,6 +48,7 @@
     title={OpenStudio SDK User Docs: Getting Started - About Measures},
     author={{National Renewable Energy Laboratory, Argonne National Laboratory, Lawrence Berkeley National Laboratory, Oak Ridge National Laboratory, Pacific Northwest National Laboratory}},
     howpublished={\url{https://nrel.github.io/OpenStudio-user-documentation/getting_started/about_measures/}},
+    year={2024},
     note={Accessed: 2024-05-28}
 }
 
@@ -53,6 +56,7 @@
     title={Decarbonizing the U.S. Economy by 2050: A National Blueprint for the Buildings Sector},
     author={{U.S. Department of Energy}},
     howpublished={\url{https://www.energy.gov/eere/articles/decarbonizing-us-economy-2050}},
+    year={2024},
     note={Accessed:2024-07-30}}
 
 @misc{skyspark,
@@ -91,8 +95,9 @@
 }
 
 @manual{energyplus,
-  title        = {EnergyPlus Energy Simulation Software},
+  title        = {EnergyPlus Energy Simulation Software 24.1.0},
   author       = {{U.S. Department of Energy}},
+  year = {2024},
   organization = {U.S. Department of Energy},
   url          = {https://energyplus.net/}
 }
@@ -111,7 +116,7 @@
 }
 
 @misc{modelica,
-  author       = {Modelica Association},
+  author       = {{Modelica Association}},
   title        = {Modelica – A Unified Object-Oriented Language for Physical Systems Modeling},
   year         = {2023},
   note         = {Modelica Specification Version 3.6},
