### 2025-09-08 #8083 @FWuellhorst
@leijerry888 Thanks for the updates.

While going through the rebuttal, one open point is
> does it make sense to include all examples and demo in the python folder? A designated demo/example would be clearer from my point of view. #rebut

I tried to see how your changes in the examples but failed to execute files under constrain/demo, either due to import errors or `'NoneType' object has no attribute 'run_workflow'` error. Could you include a clear executable example file which is ready to run? Because the official doc is still for the main, so I don't know how to run the new branch. This way, I can test your claims and changes.

Aside from that, the changes look good to me.
