### 2025-09-30 ffbcf0d
Updated CHW verifications to make them work with HW
