+++ constrain/library/chilled_water_hot_water_plant_sizing_chiller_boiler_short_cycling.py
@@ -1,7 +1,7 @@
 """
 ### Description
 
-- Chillers should be designed and controlled to prevent excessive cycling.
+- Chillers/Boilers should be designed and controlled to prevent excessive cycling.
 
 ### Code requirement
 
@@ -12,15 +12,15 @@
 
 ### Verification Approach
 
-We aim to identify the number of on/off transitions for each chiller in the system for each hour of operation. The verification passes if the number of transitions per hour does not exceed the maximum allowed cycles.
+We aim to identify the number of on/off transitions for each chiller/boiler in the system for each hour of operation. The verification passes if the number of transitions per hour does not exceed the maximum allowed cycles.
 
 ### Verification Applicability
 
-- Building Type(s): any with chilled water plant
+- Building Type(s): any with chilled/hot water plant
 - Space Type(s): N/A
-- System(s): chilled water plant
+- System(s): chilled/hot water plant
 - Climate Zone(s): any
-- Component(s): chillers
+- Component(s): chillers/boilers
 
 ### Verification Algorithm Pseudo Code
 
@@ -39,35 +39,34 @@ end
 
 ### Data requirements
 
-- status_chiller: Chiller operation status
+- status_equipment: Chiller/Boiler operation status
   - Data Value Unit: binary
-  - Data Point Affiliation: Chiller operation status
+  - Data Point Affiliation: Chiller/Boiler operation status
 
 - cycles_number_maximum: Maximum allowed number of cycles per hour
   - Data Value Unit: count
   - Data Point Affiliation: Design specification
 """
 
-from math import ceil
 import pandas as pd
 
 from constrain.checklib import RuleCheckBase
 
 
-class ChilledWaterPlantSizingChillerShortCycling(RuleCheckBase):
+class ChilledWaterHotWaterPlantSizingChillerBoilerShortCycling(RuleCheckBase):
     points = [
-        "status_chiller",
+        "status_equipment",
         "cycles_number_maximum",
     ]
 
     def verify(self):
         # Normalize the status of the chiller
-        self.df["status_chiller"] = self.df.apply(
-            lambda x: 1 if x["status_chiller"] > 0 else 0, axis=1
+        self.df["status_equipment"] = self.df.apply(
+            lambda x: 1 if x["status_equipment"] > 0 else 0, axis=1
         )
 
         # Identify transitions of operation
-        transitions = self.df["status_chiller"].diff()
+        transitions = self.df["status_equipment"].diff()
 
         # Initialization
         cycle_ends = []
