### 2025-09-30 477194e
Added system type parameter/Add more unit tests
