+++ tests/test_chilled_water_hot_water_plant_sizing_chiller_boiler_short_cycling.py
@@ -9,12 +9,13 @@ from library import *
 
 
 class TestChilledWaterHotWaterPlantSizingChillerBoilerShortCycling(unittest.TestCase):
-    def test_chilled_water_plant_sizing_chiller_short_cycling_pass(
+    def test_chilled_water_hot_water_plant_sizing_chiller_boiler_short_cycling_pass_chiller(
         self,
     ):
         points = [
             "status_equipment",
             "cycles_number_maximum",
+            "system_type",
         ]
         timestamp = [
             datetime(2023, 11, 1, 9, 0),
@@ -30,17 +31,17 @@ class TestChilledWaterHotWaterPlantSizingChillerBoilerShortCycling(unittest.Test
             datetime(2023, 11, 1, 11, 59),
         ]
         data = [
-            [0, 3],
-            [1, 3],
-            [1, 3],
-            [0, 3],
-            [1, 3],
-            [0, 3],
-            [0, 3],
-            [1, 3],
-            [1, 3],
-            [1, 3],
-            [0, 3],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
         ]
         df = pd.DataFrame(data, columns=points, index=timestamp)
         verification_obj = run_test_verification_with_data(
@@ -51,12 +52,13 @@ class TestChilledWaterHotWaterPlantSizingChillerBoilerShortCycling(unittest.Test
         binaryflag = verification_obj.check_bool()
         self.assertTrue(binaryflag)
 
-    def test_chilled_water_plant_sizing_chiller_short_cycling_fail(
+    def test_chilled_water_hot_water_plant_sizing_chiller_boiler_short_cycling_fail_chiller(
         self,
     ):
         points = [
             "status_equipment",
             "cycles_number_maximum",
+            "system_type",
         ]
         timestamp = [
             datetime(2023, 11, 1, 9, 0),
@@ -75,20 +77,112 @@ class TestChilledWaterHotWaterPlantSizingChillerBoilerShortCycling(unittest.Test
             datetime(2023, 11, 1, 11, 59),
         ]
         data = [
-            [0, 3],
-            [1, 3],
-            [1, 3],
-            [0, 3],
-            [1, 3],
-            [0, 3],
-            [0, 3],
-            [1, 3],
-            [0, 3],
-            [1, 3],
-            [0, 3],
-            [1, 3],
-            [0, 3],
-            [0, 3],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [1, 3, "Chiller"],
+            [0, 3, "Chiller"],
+            [0, 3, "Chiller"],
+        ]
+        df = pd.DataFrame(data, columns=points, index=timestamp)
+        verification_obj = run_test_verification_with_data(
+            "ChilledWaterHotWaterPlantSizingChillerBoilerShortCycling", df
+        )
+
[…]
