### 2025-09-30 2c6f741
Removed unused constant
