### 2025-10-06 7d3254f
House cleaning, add missing verification case, add unit test, fix issues with ep simulation calls, update example files to latest version of ep.
