+++ tests/test_hw_reset.py
@@ -1,0 +1,435 @@
+import datetime
+import sys
+import os
+import unittest
+
+sys.path.append("./constrain")
+import pandas as pd
+from constrain.lib_unit_test_runner import *
+from constrain.library import *
+
+
+class TestHWReset(unittest.TestCase):
+    def test_hw_reset_pass_no_flow(self):
+        """Test HWReset passes when there is no hot water flow"""
+        points = [
+            "temperature_air_outdoor",
+            "temperature_air_outdoor_max",
+            "temperature_air_outdoor_min",
+            "temperature_water_hot",
+            "flow_mass_water_hot",
+            "temperature_water_hot_setpoint_max",
+            "temperature_water_hot_setpoint_min",
+        ]
+
+        timestamp = [
+            datetime(2023, 1, 15, 11, 0, 0),
+            datetime(2023, 1, 15, 12, 0, 0),
+            datetime(2023, 1, 15, 13, 0, 0),
+            datetime(2023, 1, 15, 14, 0, 0),
+            datetime(2023, 1, 15, 15, 0, 0),
+        ]
+
+        # No flow condition - should pass regardless of temperatures
+        data = [
+            [-5, 15, -10, 45, 0.0, 80, 40],  # No flow
+            [5, 15, -10, 30, 0.0, 80, 40],  # No flow
+            [10, 15, -10, 70, 0.0, 80, 40],  # No flow
+            [-10, 15, -10, 50, 0.0, 80, 40],  # No flow
+            [15, 15, -10, 35, 0.0, 80, 40],  # No flow
+        ]
+        df = pd.DataFrame(data, columns=points, index=timestamp)
+
+        verification_obj = run_test_verification_with_data("HWReset", df)
+        binaryflag = verification_obj.check_bool()
+        self.assertTrue(binaryflag)
+
+    def test_hw_reset_pass_low_outdoor_temp(self):
+        """Test HWReset passes when outdoor temp is low and hot water temp is at max setpoint"""
+        points = [
+            "temperature_air_outdoor",
+            "temperature_air_outdoor_max",
+            "temperature_air_outdoor_min",
+            "temperature_water_hot",
+            "flow_mass_water_hot",
+            "temperature_water_hot_setpoint_max",
+            "temperature_water_hot_setpoint_min",
+        ]
+
+        timestamp = [
+            datetime(2023, 1, 15, 11, 0, 0),
+            datetime(2023, 1, 15, 12, 0, 0),
+            datetime(2023, 1, 15, 13, 0, 0),
+            datetime(2023, 1, 15, 14, 0, 0),
+            datetime(2023, 1, 15, 15, 0, 0),
+        ]
+
+        # Low outdoor temp (≤ min threshold), hot water at max setpoint
+        data = [
+            [-10, 15, -10, 80, 1.0, 80, 40],  # OAT at min, HW at max setpoint
+            [
+                -12,
+                15,
+                -10,
+                79.5,
+                1.0,
+                80,
+                40,
+            ],  # OAT below min, HW near max setpoint (within tolerance)
+            [-15, 15, -10, 80, 1.0, 80, 40],  # OAT below min, HW at max setpoint
+            [
+                -20,
+                15,
+                -10,
+                79.8,
+                1.0,
+                80,
+                40,
+            ],  # OAT well below min, HW near max setpoint
+            [
+                -8,
+                15,
+                -10,
+                80.2,
+                1.0,
+                80,
+                40,
+            ],  # OAT below min, HW slightly above max (should still pass)
+        ]
+        df = pd.DataFrame(data, columns=points, index=timestamp)
+
+        verification_obj = run_test_verification_with_data("HWReset", df)
[…]
