+++ constrain/demo/tspr_cases/tspr_cases.py
@@ -1,0 +1,161 @@
+"""The goal of this demo cases is to showcase how verification cases can be semi-automatically created for building energy simulation timeseries
+
+The 
+"""
+
+import matplotlib.pyplot as plt
+import json, glob
+
+import sys
+import os
+
+sys.path.append(os.path.join(os.path.dirname(__file__), "..", ".."))
+
+import constrain as cs
+from constrain.workflowsteps import *
+from constrain.library import *
+from constrain.libcases import *
+
+# Input/Output params
+energy_plus_simulation = True
+energy_plus_path = "/Applications/EnergyPlus-25-1-0/energyplus"
+energy_plus_idd = "/Applications/EnergyPlus-25-1-0/Energy+.idd"
+energy_plus_weather_file = (
+    "/Applications/EnergyPlus-25-1-0/WeatherData/USA_CO_Golden-NREL.724666_TMY3.epw"
+)
+energy_plus_output_file_name = "eplusout.csv"
+
+# Verification cases intialization
+verification_case_list = {"sat_reset": "SupplyAirTempReset", "chw_reset": "CHWReset"}
+verification_case_file = "./tspr_verification_cases.json"
+cases = {}
+cases["cases"] = []
+case_counter = 0
+for idf in glob.glob("./sat_reset/*.idf"):
+    if not "injected" in idf:
+        sat_case = {}
+        sat_case["no"] = case_counter
+
+        idf_json = json.load(open(idf.replace(".idf", ".epJSON")))
+
+        # Retrieve supply inlet node name of first airloop
+        airloops = idf_json["AirLoopHVAC"]
+        airloop = list(airloops.keys())[0]
+        sat_node = airloops[airloop]["supply_side_outlet_node_names"]
+        if sat_node in idf_json["NodeList"]:
+            sat_node = idf_json["NodeList"][sat_node]["nodes"][0]["node_name"]
+        # Retrieve zone design cooling temperature
+        schs = idf_json["Schedule:Day:Interval"]
+        sch = [
+            s for s in list(schs.keys()) if "_Cooling_Schedule_Summer_Design_Day" in s
+        ][0]
+        tz_coo = 99
+        for d in schs[sch]["data"]:
+            if d["value_until_time"] < tz_coo:
+                tz_coo = d["value_until_time"]
+        # Define simulation IO
+        sat_case["run_simulation"] = energy_plus_simulation
+        sat_case["simulation_IO"] = {
+            "idf": "{}".format(idf.split("\\")[-1]),
+            "idd": f"{energy_plus_idd}",
+            "weather": f"{energy_plus_weather_file}",
+            "output": f"{energy_plus_output_file_name}",
+            "ep_path": f"{energy_plus_path}",
+        }
+        sat_case["expected_result"] = "pass"
+        sat_case["verification_class"] = verification_case_list["sat_reset"]
+        sat_case["datapoints_source"] = {
+            "idf_output_variables": {
+                "temperature_air_supply_setpoint": {
+                    "subject": f"{sat_node}",
+                    "variable": "System Node Setpoint Temperature",
+                    "frequency": "detailed",
+                }
+            },
+            "parameters": {"temperature_air_zone_design_cool_setpoint": tz_coo},
+        }
+        cases["cases"].append(sat_case)
+        case_counter += 1
+json.dump(cases, open("./tspr_verification_cases.json", "w"), indent=4)
+
+for idf in glob.glob("./chw_reset/*.idf"):
+    if not "injected" in idf:
+        chw_case = {}
+        chw_case["no"] = case_counter
+
+        idf_json = json.load(open(idf.replace(".idf", ".epJSON")))
+
[…]
