### 2025-10-06 73c47e2
Deactivate simulations.
