### 2025-10-06 79af7f7
Add unit test.
