### 2025-10-06 0741e2c
Remove notebook and fix formatting and add comments.
