+++ tests/test_tspr_cases_integration.py
@@ -18,7 +18,7 @@ class TestTSPRCasesIntegration(unittest.TestCase):
         self.demo_path = project_root / "constrain" / "demo" / "tspr_cases"
         self.report_file = self.demo_path / "report_summary.md"
         self.script_file = self.demo_path / "tspr_cases.py"
-        
+
         # Clean up any existing report file
         if self.report_file.exists():
             self.report_file.unlink()
@@ -30,70 +30,79 @@ class TestTSPRCasesIntegration(unittest.TestCase):
 
     def test_tspr_cases_script_generates_passing_report(self):
         """Test that tspr_cases.py script runs successfully and generates a report where all samples pass"""
-        
+
         # Verify the script file exists
-        self.assertTrue(self.script_file.exists(), f"Script file {self.script_file} does not exist")
-        
+        self.assertTrue(
+            self.script_file.exists(), f"Script file {self.script_file} does not exist"
+        )
+
         # Change to the demo directory and run the script
         original_cwd = os.getcwd()
         try:
             os.chdir(self.demo_path)
-            
+
             # Run the tspr_cases.py script
             result = subprocess.run(
                 [sys.executable, "tspr_cases.py"],
                 capture_output=True,
                 text=True,
-                timeout=300  # 5 minute timeout
+                timeout=300,  # 5 minute timeout
             )
-            
+
             # Check that the script ran successfully
-            self.assertEqual(result.returncode, 0, 
-                           f"Script failed with return code {result.returncode}.\n"
-                           f"STDOUT: {result.stdout}\n"
-                           f"STDERR: {result.stderr}")
-            
+            self.assertEqual(
+                result.returncode,
+                0,
+                f"Script failed with return code {result.returncode}.\n"
+                f"STDOUT: {result.stdout}\n"
+                f"STDERR: {result.stderr}",
+            )
+
         finally:
             os.chdir(original_cwd)
-        
+
         # Verify that the report file was generated
-        self.assertTrue(self.report_file.exists(), 
-                       f"Report file {self.report_file} was not generated")
-        
+        self.assertTrue(
+            self.report_file.exists(),
+            f"Report file {self.report_file} was not generated",
+        )
+
         # Read and parse the report file
-        with open(self.report_file, 'r') as f:
+        with open(self.report_file, "r") as f:
             report_content = f.read()
-        
+
         # Verify the report is not empty
         self.assertGreater(len(report_content.strip()), 0, "Report file is empty")
-        
+
         # Parse the markdown table to extract verification results
-        lines = report_content.strip().split('\n')
-        
+        lines = report_content.strip().split("\n")
+
         # Find the table header and data rows
         table_started = False
         data_rows = []
-        
+
         for line in lines:
             line = line.strip()
             if not line:
                 continue
-                
+
             # Skip the header and separator lines
-            if line.startswith('| Case No.') or line.startswith('| ----'):
+            if line.startswith("| Case No.") or line.startswith("| ----"):
                 table_started = True
                 continue
-                
+
             # Process data rows
-            if table_started and line.startswith('|') and line.endswith('|'):
+            if table_started and line.startswith("|") and line.endswith("|"):
                 # Split the line by | and clean up whitespace
-                columns = [col.strip() for col in line.split('|')[1:-1]]  # Remove empty first/last elements
+                columns = [
+                    col.strip() for col in line.split("|")[1:-1]
+                ]  # Remove empty first/last elements
                 if len(columns) >= 7:  # Ensure we have all expected columns
                     data_rows.append(columns)
-        
+
         # Verify we found at least one data row
         self.assertGreater(len(data_rows), 0, "No data rows found in the report table")
-        
+
         # Check each verification case
         for i, row in enumerate(data_rows):
             case_no = row[0].strip()
[…]
