### 2026-01-18 #8083 @leijerry888
> [@leijerry888](https://github.com/leijerry888) I posted a small bug in the PR ([pnnl/ConStrain#94 (comment)](https://github.com/pnnl/ConStrain/pull/94#issuecomment-3602774412)). Otherwise, the demo is helpful.
>
> From my side, the only comments which remain are the one posted here: [#8083 (comment)](https://github.com/openjournals/joss-reviews/issues/8083#issuecomment-2909884205) - while you adressed the code-related issues, please also respond to these broader questions regarding software claims. Aside from that, I have no further review comments.

Thanks @FWuellhorst . We will respond by updating the code and paper draft.

@AdamRJensen. @lazlop also responded at https://github.com/pnnl/ConStrain/pull/94#issuecomment-3756335755
