### 2026-02-16 #8083 @AdamRJensen
I'm used to seeing grant numbers listed in the acknowledgements of funding for US national labs. Please check if the acknowledgement is correct.

Please address:
- [ ] https://github.com/pnnl/ConStrain/pull/110
- [ ] https://github.com/pnnl/ConStrain/pull/109
