### 2026-02-16 #8083 @AdamRJensen
@kyleniemeyer This submission has a non-standard [license](https://github.com/pnnl/ConStrain/blob/develop/LICENSE.md). Will you approve it?
