### 2026-02-16 #8083 @leijerry888
> > ❌ MISSING DOIs
> >
> > * 10.2172/1764632 may be a valid DOI for title: Implementation of Energy Code Controls Requirement...
> > * 10.2172/1400347 may be a valid DOI for title: Impacts of commercial building controls on energy ...
> > * 10.2172/2338089 may be a valid DOI for title: Decarbonizing the U.S. Economy by 2050: A National...
>
> [@leijerry888](https://github.com/leijerry888) Could you please check these DOIs

Will do. Thanks for the notice.
