### 2026-02-16 #8083 @kyleniemeyer
Submissions must use an [OSI-approved open source license](https://opensource.org/licenses) to be accepted at JOSS.
