### 2026-02-16 #8083 @kyleniemeyer
Additional corporate legal statements are fine, as long as they don't modify the license, but the license itself should be a standard open-source license.
