### 2026-02-17 #8083 @AdamRJensen
> Submissions must use an [OSI-approved open source license](https://opensource.org/licenses) to be accepted at JOSS.

@leijerry888 The current license is not a standard OSS license, which is required for JOSS publications. This is therefore currently a blocker to having your paper published. Please prioritize addressing this before the other issues.
