### 2026-02-18 #8083 @leijerry888
@AdamRJensen We are updating our license at https://github.com/pnnl/ConStrain/pull/111. Please let us know if the changes in this PR satisfies the requirements.
