### 2026-02-19 #8083 @AdamRJensen
> [@AdamRJensen](https://github.com/AdamRJensen) We are updating our license at [pnnl/ConStrain#111](https://github.com/pnnl/ConStrain/pull/111). Please let us know if the changes in this PR satisfies the requirements.

This is good. Please ping me once you've addressed the remaining issues.
