### 2026-03-02 #8083 @leijerry888
> > [@AdamRJensen](https://github.com/AdamRJensen) We are updating our license at [pnnl/ConStrain#111](https://github.com/pnnl/ConStrain/pull/111). Please let us know if the changes in this PR satisfies the requirements.
>
> This is good. Please ping me once you've addressed the remaining issues.

Hi @AdamRJensen, we have addressed all remaining issues. Please let us know if you have any other questions or concerns.

Jerry
