### 2026-03-17 #8083 @leijerry888
> [@leijerry888](https://github.com/leijerry888) Please address all comments in [pnnl/ConStrain#88](https://github.com/pnnl/ConStrain/issues/88) and [pnnl/ConStrain#87](https://github.com/pnnl/ConStrain/issues/87). When this is comleted, I can proceed with the editorial process.

Thanks @AdamRJensen for checking. Yes These two have been addressed and summarized in https://github.com/pnnl/ConStrain/pull/94 and paper changes merged in the paper PR https://github.com/pnnl/ConStrain/pull/33.

We are currently going through an internal review process of the final draft, please hold on for finalizing the editorial process. I will get back to you once the draft is final and ready on our side.
