### 2026-04-09 #8083 @leijerry888
Thanks @AdamRJensen for your patience. We have completed the internal review and the paper is ready for your review and processing steps. Please note that we have updated the manuscript and the reference bib on branch `joss_paper` (https://github.com/pnnl/ConStrain/pull/33) . Please let me know if you have any questions.

I believe the only thing left on our side is to merge the changes on https://github.com/pnnl/ConStrain/pull/94 and make a release. We will work on this in the coming days.
